[roll] Roll fuchsia [scripts][fuzzing] Use sys.executable instead of 'python3.8'
When monitoring a fuzzer in the background, `fx fuzz` starts a python
subprocess. Previously it was doing this with an argument of
'python3.8', which didn't reliably find the prebuilt python. Using
sys.executable always finds the interpreter the script is using.

Original-Bug: 91827
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635310
Original-Revision: b0558779069e5f5ec917bd386b87ebe20d33d3dd
GitOrigin-RevId: 4f845523faafe8d934e9a14058512f3461a496e7
Change-Id: I6abffd4bf6771735723dba32babbccfb8d9fd685