[roll] Roll fuchsia [tcpdump] Replace netdump with tcpdump for fx sniff
Replace the view option with a single wireshark option which can be used
to view a packet capture on wireshark in real-time. Note that this drops
the hex view but we do not seem to have any clients using that feature.
Note that the default view is supported by not including any option.

Original-Bug: 91943

Test:
- fx sniff ethp0003
- fx sniff -t 10 ethp0003
- fx sniff --file test ethp0003
- fx sniff --wireshark ethp0003
- fx sniff --wireshark --file test ethp0003
- fx sniff --wireshark --file test -t 10 ethp0003

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635603
Original-Revision: ee50b7433285470520dc46a434646f1844c0f283
GitOrigin-RevId: d5da969050c9f9451598335397492014946e5db0
Change-Id: Ic9a054d0e57ed7c65a5a9fc4cbdaafc159cd1db4