[roll] Roll fuchsia [assembly_util] remove lifetimes from traits and implement on &mut T
I think this is a bit simpler because you don't have to thread the
lifetimes around everywhere and I think it makes more conceptual sense
to have the trait not be generic over the lifetime and leave the
lifetime with the type.

As far as I can tell this doesn't matter to callers because method
dispatch still finds this trait no problem (I didn't have to change the
tests).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635824
Original-Revision: 6741aa57ff11a4d30bd5c9bc0698fc06ed1c04cd
GitOrigin-RevId: 15aacf06f4bda2883caf463eb8a97da351980626
Change-Id: I0689dd8796519d65d51845231197ef93e618bdff