[roll] Roll fuchsia [graphics][compute][spinel2] Fix interactive test and deps reentrancy
This CL fixes a reentrancy issue in the `deps.c` dispatch scheduler.

All path, raster, composition and styling objects invoke the
dependency (deps) module to submit dispatches to the device.

A completion action is executed when the dispatch is finished.

The path and raster completion actions were invoking path and raster
handle reclamation which in turn invoked the dependency module.

This kind of reentrancy was supported but broke the goal of being able
to serialize all dispatches with a single timeline semaphore --
something that was useful for debugging.

The fix involved:

 * Delay execution of all completion actions until after an
   "immediate" semaphore has been acquired.

 * Update the handle pool reclamation routines to support reentrancy.

 * Updates to the raster builder to track dispatch state more
   precisely similar to the handle pool.  This wasn't an issue but the
   rest of the Spinel dispatches will mirror this improvement since it
   improves debug'ability.

Additionally, the host interactive test on XCB assumes BGRA ordering.
For now, and as a short term solution, platform-specific compilation
ensures that the renderer properly swizzles the components from RGBA
to BGRA.

Finally, the host interactive test running on NVIDIA 470.86 can hang
waiting for a semaphore to be signalled.  This appears to be resolved
in the 495.xx driver.

This will be further tested the against the Intel/ANV Spinel target and driver.

Supported targets:

 * NVIDIA sm_35+

The previous list of targets will be added in a following CL.

Testability:

 * Build: fx set core.chromebook-x64 --with src/graphics/lib/compute/spinel2/platforms/vk/tests
 * GTest: out/core.chromebook-x64/host_x64/exe.unstripped/spinel-vk-test
 * Interactive: out/core.chromebook-x64/host_x64/exe.unstripped/spinel-vk-bench -f <svg>

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/616643
Original-Revision: 839393a57071f768b9d6359148c12470c5d9057d
GitOrigin-RevId: 3986d7dd45db7e151197bcec667c58e3ffdb860f
Change-Id: Ie43f910423f8fd625417acb31fc768e35a6b5b16