[roll] Roll fuchsia [ffx] Clean up error messages for deserialization
As we've been developing ffx emu, there hae been a lot of changes to the
serialized configuration along the way. Each time a state file is left
on the disk and the configs change, we hit errors on start-up and
shutdown about some field missing, but no indication of how to fix it.

This CL adds text for these serialization errors to indicate 1) where
they came from, and 2) what to do about them. We modify the output for
both the start, shutdown, and show subcommands.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635982
Original-Revision: 12a7cf695baa5a58c98085fecd78bf14580450d5
GitOrigin-RevId: 4e240afbcc52c4479a0f399ae9d8536a91d20b1f
Change-Id: Iabbff324b11ecf936e2a767056fb59e641387908