[roll] Roll infra/recipes [git] Add helper to cherry-pick a CL
In special circumstances (see bug), it can be desirable to cherry-pick a
CL without rebasing. The cherry_pick_cl() function resolves a
GerritChange's fetch ref, fetches it, and cherry-picks it in the current
working directory.

This is identical behavior to using the triple-dot menu in Gerrit and
selecting the Download Patch -> Cherry-pick commands. It is still up to
the caller to run these commands in the appropriate working directory,
and so if used inside a Jiri checkout, the caller must resolve the
working directory of the CL's project first (using api.jiri.project).

Original-Bug: 87199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/636283
Original-Revision: 6d2bd74719a489357b3e3188842ca7a8c94e2720
GitOrigin-RevId: 306a2025b68b061375d1db10f8624bb52f4212bd
Change-Id: I2f6417a77f3ca053e83e3b6829630e178f4ba093