[roll] Roll fuchsia [ffx] Add graph orientation flag.
This lets the user change the orientation of the graph to left-to-right via `ffx component graph --orientation=left_to_right`.

This change also changes the default orientation from left-to-right
to Graphviz's default of top-to-bottom.
This tends to look better although it takes up more room.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632601
Original-Revision: 32672f6ecbfdc02fe858930f6d63918399916381
GitOrigin-RevId: ea6a8755aca895570ad5598704f48fe3f23f6514
Change-Id: I665594392124687d14ac0a99e484afa38db20ddf