[roll] Roll fuchsia [fshost] start fshost rewrite
This adds an experimental rust version of fshost. Currently the rust
version implements fuchsia.fshost.Admin for shutting down, and that's
it.

An integration test harness is added, using RealmBuilder, which tests
behaviors against both the rust and the cpp version of fshost to build
confidence in the replication of these behaviors in the rewrite.

Original-Bug: 87511
Test: fx test fshost_integration_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598882
Original-Revision: b2e65b3582d33aeb62861cf9744953dbd66dc6d4
GitOrigin-RevId: 0708ad29ee652d2333fc9a86d4a390f245a7d209
Change-Id: I744eeceb72e83da39618ad4028241b31408e9915