[roll] Roll fuchsia [component_manager] Hub permits absent parent at child discovery
The CL returns Ok, instead of an error, if a parent component is not
present in the `instance_map` at the time its child is discovered.

Additionally, the `Hub::add_instance_if_necessary` call was refactored
to return early if `None` is returned (ie the target component is
already in the instance map) to reduce drift with nested conditionals.

The reason for the CL is to address the scenario where a purge event for
the parent has removed it from the instance map before a discovered
event for its child has been processed. A TODO is added for
fxbug.dev/89503 to investigate this event ordering scenario further.

Original-Bug: 89503
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635825
Original-Revision: 2b65bf4ed3b563336d16f00a3ade7b6b403e262d
GitOrigin-RevId: b754ef4013802f3e0267e716bd50395130ed4780
Change-Id: I0a31095130496b1d3ebd2c58e185a94b296297bd