[roll] Roll fuchsia [fxfs] Eagerly reserve metadata space
The global metadata reservation should always be created before any
transactions are attempted, and it is an invariant that we should always
have enough space for that reservation.

Because this reservation is lazily initialized, it is possible for it to
be initialized after some other transaction has stolen space from the
allocator, e.g. if the first transaction requires its own reservation.

Fix this to eagerly reserve the space when we start up.

Also add tests for low space conditions to fs_test, since that is what
triggers this bug.

Original-Fixed: 91475
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631772
Original-Revision: c9bf95133e77782dc168f9b6e583cbfa4c06640a
GitOrigin-RevId: b1139dad2cbf584babc34dd64ae00f696836f0e2
Change-Id: I3f7a648ecc79686855ffc780e9939f9415071557