[roll] Roll fuchsia [preflight] Add DevRel team as OWNERS
As the team looks to resolve some bugs and feature requests in
preflight. It makes sense to have DevRel claim ownership of this plugin
and experience especially as it is critical for on-boarding of new
fuchsia developers (as documented in fuchsia.dev).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637411
Original-Revision: e6999686f9b2c88935846e4e02c960901c457c24
GitOrigin-RevId: 0183533b7ad59a55b70e27a7d084a76dc9e0a0f9
Change-Id: I35ab1ade158c963b0caf51e249e3bc97e3bfb1f3