[roll] Roll fuchsia [blobfs] remove fuchsia.blobfs c fidl bindings
Only a couple of places imported them and those places didn't even use
them. Remove the imports and the annotation that generates the bindings.

This also adds discoverable to the fuchsia.blobfs.Blobfs protocol. After
blobfs is a component, this protocol is going to be placed in the
outgoing directory instead of being multiplexed over directory
connections, so I added it in since we will eventually need it and we
are modifying it anyway.

Original-Bug: 90556
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633774
Original-Revision: 2997253cc29aaa76c1d05db4281348aadfe4cb22
GitOrigin-RevId: 6f69acbd596e7152bebba8d3f9c027807b278fb8
Change-Id: I410b271dad42a10fe2b45a5792eabd99e4749757