[roll] Roll fuchsia [assembly] add pkgfs_package_label param to assemble_system()
Allow callers of assemble_system() to choose the pkgfs implementation that it
inserted into the images when a pkgfs implementation is needed.

This addresses an issue of duplicate pkgfs packages being added to the
images when constructing certain tests.

Original-Bug: 92037
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636122
Original-Revision: ce11ca4b54bcaa8a0605abdc598b53fe0289bcad
GitOrigin-RevId: b06a054a078c4b7424bdb75eb69b47cd8115715e
Change-Id: I56af1ce88a4c6d1347de3bf27e4a31b825adc255