[roll] Roll fuchsia [wayland] Remove ViewProducer protocol
This protocol is no longer used, is not part of the Fuchsia
wayland API roadmap, and has a number of flaws that are hard
to fix (e.g. race between apps during startup).

This removes the protocol and runner support from the Linux
runner componenet. The Linux runner still implements the
LinuxManager protocol, which provide and reliable way to
monitor Linux container startup and can be extended to
also support application launching in the future.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636822
Original-Revision: 4645274c75d65554768b425496af75caaaf4ff81
GitOrigin-RevId: 31294b128ef08821706a6b7d3fd4430f741b1065
Change-Id: I98ead95c1fff611e3ea449e7b833cd2c0c4d6bb1