[roll] Roll infra/recipes [spec] Add test_on_gce to test spec
Recipe properties are preferred over module properties for things that
configure the behavior of a specific recipe. The `experimental` module
is a bit of an antipattern because it encourages the latter, so I'm
planning on removing it in favor of instead using temporary top-level
recipe properties with a more limited scope.

`test_on_gce` is the only experiment still in active use, so I'm moving
it into a spec field instead.

Plumbing around the `test_on_gce` value is a bit annoying but I have
plans to reduce plumbing required when adding a new testing_requests
parameter, so hopefully this will become less of a reason to use the
`experimental` module in the future.

Original-Bug: 88439
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/637354
Original-Revision: abe6e471fc13c4de0a7320f37da482d2c7664ef8
GitOrigin-RevId: d145610880105b19bc1431984ad5d40c78cad41e
Change-Id: I44514c1d59ee924fb3b7e07fb975ac9d63871872