[roll] Roll fuchsia [starnix] Parse BufferCollectionToken from ImageFile
ImageFile also contains a BufferCollectionImportToken that needs to be
sent from the client to the wayland bridge. This change adds the code
that parses the token out of ImageFile, and adds an error log when the
file type is invalid.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637581
Original-Revision: dcd1e2f687ec3312e6606aba8e087dbb9bb9a7d2
GitOrigin-RevId: 52b3eb9ad58b5e2ce7213f720f4f7661d08632f9
Change-Id: I50787fdaef69736bd1f61a687096d269f7ee36bc