[roll] Roll fuchsia [elf_test_runner] Remove ordering assumptions
Removes some incorrect assumptions about the order in which events are
received. The order in which stdout/stderr events are received depends
on when the test writes to the handles, and when the test reads off the
socket.

Multiply: elf_test_runner_integration_test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637494
Original-Revision: 8f61f73625ca662a3514f9d196f19f213e3bfbee
GitOrigin-RevId: 7832746fa769509a1f3c1d3796fc4a255eff9e8e
Change-Id: If9c0c08511d40dfd4d28d4882952e4125bc65518