[roll] Roll fuchsia [display][rust] Mechanism to apply a layer configuration
Introduced a mechanism to declaratively describe a display layer
configuration and apply it to the display driver with validation. Not
all cursor and primary layer parameters are currently supported (notably
transforms, blending, and fences) but sufficient support is there to
build a primary layer with a shared image.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611802
Original-Revision: 174040486dcbb3ad01e41520a6ab35596d38aa8c
GitOrigin-RevId: e9d447b1c46033134e9a5325118a5614b6b87697
Change-Id: I99c30dc3cabcba7165658ed39569410c01a79a6a