[roll] Roll fuchsia [camera][gym] Do not stop on PEER_CLOSED (buf col)
This CL is a re-land of fxr/629444:

Continue on BufferCollection PEER_CLOSED "error" because sysmem could
simply be reflecting a PEER_CLOSED by another client, and if so, then
this is a case where there are no real errors but merely signaling a
tear down.

The new assumption is to tear down the BufferCollection as camera-gym
knows it. In this code path, there is also an allowance for potential
race condition where camera-gym might a tear down might have occurred
already.

All errors for BufferCollection are now treated this way.

Original-Bug: b/209523797
Test: camera-gym run consistently >10 hours without camera-gym exit.
      Verify that, upon disconnect, the BufferCollection unbound case
      executes.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637416
Original-Revision: a7de8055aaad1ac15d48fb5fecb8e45266401a0e
GitOrigin-RevId: ef08d14470e311583e52448a280a944685ba466d
Change-Id: I5652d1efeeaabf78b080e309c7b8857bd448e8fb