[roll] Roll infra/recipes [crashpad_roller] Update mini_chromium project name
It was renamed in http://tqr/469615. It's not great that the roller
hardcodes the Jiri project name, but probably not worth fixing now
because the recipe will need to be rewritten for submodules, and it's a
one-off recipe anyway.

Original-Bug: 91441
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/638281
Original-Revision: 43fd6a6f975317d76dbf4de83fd8d03ae3536aa8
GitOrigin-RevId: 9a77eb9a887573c8592e17e18a56e14522a3091f
Change-Id: I11054560f1f0bdb35bb9c52ad5ce62bf3c324827