[roll] Roll infra/recipes [recipe_tools_roller] Roll packages to the oldest tag
Subsequent attempts to roll a package will not be guaranteed to produce
the same commit diff as long as the "latest" ref still points to the
same instance (assuming no deletion of tags).

Previously the diff would change if a newer tag was added to the package
instance between roll attempts, so the roller would create two separate
CLs instead of restoring the first CL if it failed CQ. This prevented
the first CL's CQ results from being reused and made it falsely appear
as if the two CLs rolled the tool to different versions.

For example, fxrev.dev/638162 and fxrev.dev/638386 actually roll the
gerrit tool to the same instance even though they use different tags to
refer to that instance.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/638562
Original-Revision: 686b49c2620794b341428f9bd234ffb1e3c120fc
GitOrigin-RevId: 0361b6f6c98d11c7b4f88e9facb8479f70c4fcd9
Change-Id: I30c216fafa5e08e70ecdfa0c0e12fea8b4ecef3d