[roll] Roll cobalt Store the complete current system profile when logging an event.
This ensures that the event is recorded with the SystemProfile that was
current at the time the event occurred. If we used the current system
profile when the event was aggregated, it could be different, if the
event was cached in the undated_event_manager for a time.

Original-Bug: 91520
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/636882
Original-Revision: 730e33f8dc922053503d0921a457b7cfd5c12d71
GitOrigin-RevId: 64bc15facf4e6ce02238689518a46b47139b39cd
Change-Id: I31e15445ce14e895a764ccc608a0fe6503936ee0