[roll] Roll fuchsia [fidl][cpp] Initial natural API test
This is the start of tests for the natural API bindings. So far it just
contains a copy of the wire API tests for enums and bits with the symbol
names changed - since those are the same types but in a different
namespace.

This also moves the fidl.llcpp.types.test FIDL library definition out of
the wire_types subdirectory since it's used for more.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635537
Original-Revision: 4c10e7ed1e71b1c81a4186e254c25af99df6a7da
GitOrigin-RevId: 83073a1394d03c744efa9043c0ace84d1c16d428
Change-Id: I6f207da4543d5c7ddc9317b70d49925072f1dc08