[roll] Roll fuchsia [input][tests] Create a (WIP) workstation test realm
This is a static definition of a workstation test realm. It can be used
from within a RealmBuilder to create a sub-realm that replicates the
static workstation setup for tests.

Note that, sadly, the RealmBuilder seemingly does not allow loading CFv1
components from within such a manifest file, so a number of components
had to be wrapped into a dedicated runner. Since this is expected to be a
one-time setup, it should not complicate the reasoning about the realm
very much.

Original-Bug: 92183
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638675
Original-Revision: e8dfbe0cf76780d62464700f742ad2c75a0c2d49
GitOrigin-RevId: 3a59c2fcc9bd04483de1cf8168c6f549af1d54e7
Change-Id: Ic0af5a143f3ba75c60116ad2be6dd7545708bd9e