[roll] Roll fuchsia [ffx] Fix the --debugger flag
Code review and testing brought up a problem with how we're handling the
debugger flag in ffx emu. When executing `ffx emy start --debugger`, we
expect the emulator process to pause before booting Fuchsia so that the
user can attach a debugging process like gdb first. Instead, the
launcher process would pause, but the emulator would just start up as
normal. If the user added the --console or --monitor flag, debugging
would be ignored.

This CL adds the appropriate -wait-for-debugger flag to the emulator
flag template, predicated on the debugger flag of course, so the
emulator will actually pause like it's supposed to. We also remove the
code that was causing the launcher to wait for the emulator to return,
as this was a carryover from device_launcher and doesn't fit our
existing execution model anymore.

Now, the launcher process will exit once the emulator is running (or
branch to the console/monitor, if needed), and the emulator process will
pause, as expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638724
Original-Revision: 19ed503886f25c48845110ca7c1431e5785c576d
GitOrigin-RevId: 5f9188bbb97a21c150490face8a5fd5edb12a257
Change-Id: I6b0675e958981ce933d53d47d38f15e611e02593