[roll] Roll fuchsia [kernel][clocks] Allow imprecise rate ratios when applying rate corrections.
When applying rate corrections, permit the ratio which converts
directly from ticks to synthetic time to be imprecise.  Essentially
what this means is that if the ratio cannot fit into a 32/32 bit
ratio, bits from the ratio will be discarded until it can.

Prior to this change, inability to store this ratio in a 32/32 bit
ratio could result in a kernel panic.  If the reference ticks -> clock
monotonic rate ratio contained a lot unusual prime factors, and the
rate correction being applied by the user also had a lot of unusual
prime factors, the net result was a ratio which would have required a
64/64 bit ratio to store precisely which was triggering a panic.
While it is never OK to allow user code to trigger a panic in the
kernel, it should be OK to allow degraded precision in this
conversion, as the users of these clocks are typically actively
maintaining the clock transformations anyway.

Tests: existing unit tests.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637679
Original-Revision: 868c6f08da4eaac8cb684f35b67c7d67da0e198f
GitOrigin-RevId: 1273c737113d79674f992731bf068f7b7f8ffe86
Change-Id: Icbe7d90cfef0d7fc46f43f1fb055035d0b695c73