[roll] Roll fuchsia [fidl][cpp] Natural event sending
Adds |fidl::SendEvent| for sending events with natural types.

It uses similar structure as |fidl::WireSendEvent|, but is much simpler
because there are no caller-allocated flavors.

The return type is |fitx::result<fidl::Error>|, which is the same as
that of one-way client calls.

Now we have everything that's needed for a pair of end-to-end async
clients and servers.

Added a test in the wire_interop_tests.

Original-Bug: 60240

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638203
Original-Revision: 955496875a85acc6339a82c7e2a11f3968d88ce1
GitOrigin-RevId: f102fbbbfa9e7516447e2210adc9a5d696a093a1
Change-Id: I345b13dd971603b0224ccdf3136ed2fe57f96ea4