[roll] Roll fuchsia [fdf][mock-ddk] Allow device_add to take nullptr
Currently, device_add_from_driver allows the |out|
argument to be null, in the situation where a driver
does not need to maintain a pointer to the device added.

This adds that behavior to the mock-ddk and the fake-ddk.

Test: Included.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638679
Original-Revision: 5d757756ecb4d2fe2a0913a53d6bde6e15f01252
GitOrigin-RevId: 2fd1754bce535fad2fa906b9cd3bb423158bfb47
Change-Id: I834ef89c63d04646a47aed5977a8135a969a855c