[roll] Roll experiences [ermine] Add logout button
This change adds a button to logout of the ermine shell. When oobe is
enabled, the user is sent to the login screen. When it is disabled, the
user is logged out and then logged back in with an empty password.

The buttons are updated to icon buttons to confirm to latest ux design.
- Adds a helper dart extension to add tooltip to any widget.

Original-Bug: 92301
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/638981
Original-Revision: 0a0708923a20d1f47e0a9f666cdabc1a4709fc92
GitOrigin-RevId: 1b4b539244c3bed19e493516137971dbfc91d96d
Change-Id: I3822c5544f7439611a2840951fc61b4fc7802a1b