[roll] Roll fuchsia [system-update-checker] Check for empty target channels
In https://fxrev.dev/627872, we switched system-update-checker to use an
empty string if we didn't find the target channel in vbmeta.
Unfortunately this caused us to try to ota with `fuchsia-pkg:///update`.
This patch checks for this condition when trying to construct the update
url, and returns `None` if the target channel is empty. This causes us
to fall back to the `fuchsia-pkg://fuchsia.com/update` url.

Original-Fixed: 91886
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635373
Original-Revision: ee870bb57a3860c82345e27e690bd42e85c9c7e1
GitOrigin-RevId: 0b13684c2b8624622292074ae94fdae527792f5c
Change-Id: I64295ae084d58c583cbb27994b543a540ca3d925