[roll] Roll infra/recipes [cache] Make cache purges atomic
I believe this will fix some `jiri update` failures that happen because
the cache is corrupted from a previous build, but the guard file doesn't
exist because the previous build got cancelled in the middle of purging
the cache, having cleaning up the guard file early on because
`api.file.rmcontents()` is not atomic and removes files in a
nondeterministic order.

For example, ci.chromium.org/b/8823804247104332497 failed due to a
corrupted cache despite the guard file not existing, likely because the
previous build on the same bot (ci.chromium.org/b/8823804666225823969)
got canceled in the middle of purging a corrupted cache.

Now we'll delete the entire cache directory in an atomic way, which also
ensures that Swarming won't consider a purged cache to still be warm (as
was the case when we just removed the cache's contents without deleting
the directory), which messes up warm cache waiting and incremental build
metrics.

Original-Bug: 87277
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/639221
Original-Revision: 7f1272d0df66d8ace98b410a66ddf43d99e0c0c1
GitOrigin-RevId: 60faeeb1a6814b4a80705446a82d58a965859f92
Change-Id: I618f52f48881f0c9a2cdabbbd5a6bc465c89a924