[roll] Roll fuchsia [minfs] quieter stat when not on fvm
When not on fvm this warning prints on every stat, which in one of the
unlink tests means it prints on the order of 70k times. Make the message
debug, and also turns up the sleep in the unlink test so we don't call
it as often while waiting for the blocks to become free.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638841
Original-Revision: 40622c322e98216add15dc852d10f5e9507430b1
GitOrigin-RevId: ec9d0b0c57ef663f52945f0ad17da238e3742f49
Change-Id: I3104e588f32c58179bd19f2883c35b7157faf479