[roll] Roll fuchsia [ffx] Create Http Repository
This patch introduces the notion of a http repository, where the ffx
repository server will proxy requests to the http repository. This allows
requests to go through the ffx tunnel in case a device cannot talk directly
to the internet.

As part of this change, we:

* Move the construction of the `Repository` types to the repo daemon
  service. This allows us to share the hyper client between http
  repositories, which allows us to transparently reuse connections.
* Initialize trust with 1.root.json, rather than root.json.
* Changes pkg::repository::package_download to use the new http repository.
* Uses RepoBuilder to construct test metadata to make sure this works.

Original-Fixed: 75396
Original-Fixed: 85213
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637083
Original-Revision: 28df06bea59021da020316d50bbcd3d82f3c0f85
GitOrigin-RevId: ef5278737ecc79f8a323b0bb94bdda9e94c10a9b
Change-Id: I23e1e1579cf56bf1b35b1acd6ae6fb43b849dd10