[roll] Roll fuchsia [assembly] Use PackageBuilder to make packages in assembly
Switches the creation of the 'base' and 'update' packages over to using
the PackageBuilder.

Adds the ability to add files created from passed-in content to the
builder.

Adds the ability to set the ABI revision to a package.

Removes the consolidated create_meta_package_file() fn from the
utils lib, as it's now part of the PackageBuilder.

Original-Bug: 91743

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/637664
Original-Revision: f8a89b37a7ee74d2fce6a08038b514be44b457f4
GitOrigin-RevId: 128bb125f62bc6e1fff21b3d3293e1fd88fcb615
Change-Id: I58d2c1b0c94e42d95524d9697647455e8632f511