[roll] Roll experiences [ermine] When dismissing views, move focus to previous view instead of next
This is important for Chromium and Linux apps where it is
confusing to the user when focused is moved to the next
view instead of what had focus previously.

Test: Launch two terminals and one spinning square, close
spinning square, focus is returned to the last focused
terminal instead of the first.
Original-Bug: 91433

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/639601
Original-Revision: fd8a6d507a0be41ffae8bd2d329b285a26d8c8e3
GitOrigin-RevId: c0e85f6bcbfd5aa4f68e962e14fa009a416b85f1
Change-Id: I7e1f972e2875b097a94e1468a89d3414de340a6f