[roll] Roll fuchsia [fidl] Refactor header allocation out of payload coding tables
We'd like to move to a regime where FIDL request/response payloads and
their equivalent non-payload types are entirely interchangeable.  More
concretely, the following two structs should generate identically
structured coding tables:

```fidl
type MyStruct = struct {
  a bool;
  b bool;
  c bool;
};

protocol MyProtocol {
  MyMethod(struct {
    a bool;
    b bool;
    c bool;
  });
}
```

In C-like bindings, there are two differences in transactional and
non-transactional type definitions like the ones above: 1) trailing
alignment padding is always included for transacational types but never
for standalone types, and 2) transactional types have space reserved for
their header, both in their overall `size` properties and in the offsets
of each of their members, whereas standalone types do not. For the
example above, `MyStructMessageTable` would have a `size_{v1|v2}` of 3
bytes and `offsets` of 0, 1, 2 respectively, whereas
`MyProtocolMyMethodRequestMessageTable` would have a `size_{v1|v2}` of
24 bytes (aligning the object, then adding 16 bytes for the header) and
`offsets` of 16, 17, 18.

This change resolves issue #2 mentioned above by removing this 16-byte
size/offset shift, and modifying the encoder/decoder logics to apply it
themselves when necessary.  Thus, as of this change, the example above
generates a `size_{v1|v2}` of 8 bytes (still undesirably aligned, but
now no longer adjusted for the header) and `offsets` of 0, 1, 2, making
it more congruent with the standalone type.

Removing the header allocation from the size and offset values in the
coding tables changes the assumptions that consuming code may make about
any `fidl_type_t` that it receives, as it is now no longer necessarily
true that the size of that type and the size of the message or buffer it
is contained in are equal. Instead, the type may or may not be in a
buffer containing `sizeof(fidl_message_header_t)` of leading bytes.

The solution implemented by this change is to make such consuming
functions context-aware by adding the `has_header` boolean argument,
used to indicate whether or not the buffer containing the type has a
leading 16-byte header that needs to be accounted for. This is useful,
for example, in the case of `EncodeIovecEtc`, which must ensure that the
first generated iovec buffer leaves extra leading space (when
`has_header == true`) for the header that is otherwise not described by
the passed in `fidl_type_t`. Care has been taken to modify as few
SDK-facing function signatures as possible when introducing this new
boolean parameter.

Test: fidldev test all
Original-Bug: 88343
Original-Bug: 88344

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621953
Original-Revision: 70055fbed13f35d2116883214a5dbc7a78e5cf4e
GitOrigin-RevId: 374e14fda42fd1f6816f4a3c02a062197463e763
Change-Id: I219787aacee378dbbf69a0f2a62d4d4ee4e5d432