[roll] Roll fuchsia [netemul] Define and parse network config
Until the component manifest schema supports it, the network
configuration for a netemul runner test will be specified in a separate
JSON file bundled with the test package.

Also, this CL does not introduce the actual configuration logic that
will exercise the netemul sandbox and the configurable netstack; it just
introduces the configuration schema and reads and validates it in the
runner.

Original-Fixed: 92175
Test: fx test netemul-runner-unit-tests
Test: fx test netemul-runner-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638601
Original-Revision: 1fd47b7fd1a049bab7b6b7e9a7b7c75ba9b88e72
GitOrigin-RevId: 036ea658581befacfe74e6c357eb93c6f46a6833
Change-Id: Iec7bce3a59d3310f92d028f615f0d3f241f1ae88