[roll] Roll fuchsia [virtio] Barrier after enabling interrupts
There is a race between the driver and the device where when the driver
re-enables interrupts it must check if it races with the device just
enqueueing a new descriptor. Unfortunately this race check is itself
racy since the load of idx can be reordered before the store to the
flags.

This commit fixes this by adding a memory barrier to ensure the load
and store are not reordered.

This change adds a sequentially consistent barrier

Note this can be imagined as two threads performing the following operations
 Thread 1:
  1. flags.store(0)
  2. x = idx.load()
Thread 2:
  1. y = flags.load()
  2. idx.store(new_idx)
Without a seq_cst barrier between statements 1 and 2 it is possible to
end up with both x != new_idx and y != 0, thus an interrupt is not
generated and the driver fails to see the returned descriptor.
Both the driver and device require a barrier.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639013
Original-Revision: 429b463d62163969542fe7e0fec3e570653628bd
GitOrigin-RevId: edd7dfc9aee89461198a5d7a951d0d4eb4348eb3
Change-Id: If3604dd723faf6e00182c2226bf366c00ded79ff