[roll] Roll fuchsia [accountctl] Print names for each account on the device.
If the use different names in different account creation flows this will
be a useful debugging aid. If not then it will still be useful at some
point in the future.

Note that account-ids doesn't make much sense as a subcommand if it
does both IDs and names so I changed to "list". I'm enjoying the fact
that thats easier to type!

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638726
Original-Revision: 227b8dd675b0cf754f7ae98c29cb76db7b29fd30
GitOrigin-RevId: 2343517df9e5bf871589cd6ab4075f48adcce538
Change-Id: I3e4d0d413d8455a7e3dd3a18d9e5cd5c496c5585