[roll] Roll fuchsia [f2fs] Prepare for FileCache
This CL introduces Page class to represent a block as
a combination of vnode, index, and vmo before implementing
FileCache class. Once a Page is allocated, f2fs guarantees
that the vmo keeps its mapping (vaddr) and the committed page
until there is no reference to it. To track the status of each Page,
Page class uses some flags such as dirty, uptodate, writeback,
mapped, alloc, and locked.

Page::index_ indicates different information according to
the type of Page::vnode_ such as file, node, and meta vnodes.
For file vnodes, it has file offset that the Page represents.
For node vnodes, it indicates the node id. For meta vnode,
it points to the block address to which the metadata is written.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/626193
Original-Revision: 7fc2f493904a6375c02911f2206bf3758a97e252
GitOrigin-RevId: df6fb237ff9d334f66c990eb68885dc682690778
Change-Id: Ib7abde502076a4b7e23b8936a51fbfd5cd365158