[roll] Roll fuchsia [utest] Reland "Move object property syscall tests into core-tests"
This is a reland of fee712ea318f98d89f8cd58d9060d415821e7137.
The test instability is introduced by the old property test's
busy-wait thread consuming a whole CPU for the rest of the
standalone test.  That is resolved by the revamp of that test
code in a follow-on change that will land together with this
reland.

Original change's description:
> [utest] Move object property syscall tests into core-tests
>
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638639

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639575
Original-Revision: 9192a0f160ae7ed5396082e4766de00a03bde291
GitOrigin-RevId: 76b8d84597bbcf6b496a0f519da9add09df190db
Change-Id: If8745e75bc4778e6593ee6b29326d2b5b068f1ca