[roll] Roll fuchsia [kernel] Allow ZX_PROP_REGISTER_{FS,GS} in zx_object_get_property, improve test
This adds a much more thorough test for the behavior related to
the %fs.base and %gs.base registers.  Partly to facilitate this
test, and partly just to make ZX_PROP_REGISTER_{FS,GS} a more
complete software substitute for the newer CPU instructions,
zx_object_get_property now accepts these properties just as
zx_object_set_property does, so it can provide a software
replacement for rdfsbase/rdgsbase as well as wrfsbase/wrgsbase.

The new test exercises the CPU instructions as well as the system
calls, so it can verify that they interact consistently.  To test
both with and without the CPU instructions available, QEMU (with
or without KVM) can be passed the `-cpu -fsgsbase` switch when
running on hardware that does actually support it to emulate
hardware that does not.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639371
Original-Revision: bcb396aec477d77cc8140489c13a10ba5e3339fe
GitOrigin-RevId: 5f8fcd5abedca5b82bbd6c17ad994e38baaa11eb
Change-Id: I3356954834103ef746815db72c92e6ce415c770f