[roll] Roll fuchsia [core_size_limits] Set absolute limits
The size limits were set to an arbitrarily high limit before. This
change modifies the size limits to be high, but within the realm of
reason so that the numbers won't look ridiculous once we show them in
the binary size table.

The size limits were determined by taking max(core.arm64, core.x64) for
each component's current byte consumption, multiplied by 1.5, so there
is plenty of room to grow. The intent remains that we do not wish to
place a cap on the core product's components, only that we want to use
the creep limits to make organic growth more intentional.

The next step to this will be to add core_size_limits.arm64 and
core_size_limits.x64 to the binary size table in Gerrit, but as FYI
only.

Original-Bug: 83734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639580
Original-Revision: c9f5a29bf1f72fd9d67d2ea4d50c1f9528d4fcf7
GitOrigin-RevId: b91bc236484721e40f6f3a344289e774d9800efc
Change-Id: I59e2b2a04e44b4e4e4da97f0dc017595cd5c0a21