[roll] Roll fuchsia [build] Allow actions to specify ignore prefixes for action_tracer
This allows actions to specify a list of prefixes which the action tracer
will ignore.  This is useful for mimicking directory-glob semantics where
an action needs to recursively delete an entire directory tree in the
outdir, but on an incremental rebuild, the contents of the directory tree
may be different.

Original-Fixed: 92503

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639577
Original-Revision: 83ee690ad5f7c2eef44f8f1e2b96c5ebf3c9256e
GitOrigin-RevId: e87590ddf339c908fe5e63fbaa97247373960dd5
Change-Id: I282535e2b79c442e2dbe72b90a1254d12f535c7b