[roll] Roll fuchsia [moniker] Rename PartialAbsoluteMoniker to AbsoluteMoniker
`PartialAbsoluteMoniker` (the moniker representation that does
not include instance ids) is renamed to `AbsoluteMoniker`. Struct fields
and variables named `partial_abs_moniker` are named to `abs_moniker`.

This CL is a rename only.

The reason for this CL is to normalize the use of formerly named
`PartialAbsoluteMoniker` as the only moniker representation available
for future development.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639045
Original-Revision: 5244fba9a5b6c39df81daa9627c578079ad19071
GitOrigin-RevId: 0bef1c013da1e581baa200f0f575bb2532d2d71d
Change-Id: I4382554d1a6c99edfe065ca22c00296ebd07b019