[roll] Roll fuchsia [moniker] Rename ChildMoniker to InstancedChildMoniker
`ChildMoniker` (the moniker representation inclusive of instance ids) is
renamed to `InstancedChildMoniker`. The `ComponentInstanceInterface`
trait function `child_moniker` and variables that point to an
`InstancedChildMoniker` are renamed to `instanced_child_moniker`.

This CL is a rename only and follows the conventions of
`InstancedAbsoluteMoniker`.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639046
Original-Revision: b6c578dd0b9ce92d2534019c72ac279aef088b4d
GitOrigin-RevId: 007a23294283875496165a60e04ee0f141a24835
Change-Id: Ibad6fb9f7b794b090304f9a07ad6dacb011d7162