[roll] Roll fuchsia [moniker] Rename RelativeMoniker to InstancedRelativeMoniker
`RelativeMoniker` (the moniker representation that includes instance
ids) is renamed to `InstancedRelativeMoniker`. Variables that point to
an `InstancedRelativeMoniker` are renamed to
`instanced_relative_moniker`.

This CL is a rename only and follows the conventions of
`InstancedAbsoluteMoniker`.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639048
Original-Revision: f2ee9d34e6834244d840f0a4efb0380b4d39a9e6
GitOrigin-RevId: d7a408d97bc7cf034c661a6b9670bc68cdb7a833
Change-Id: Ie0e0943223b280816607a1227af923b8edf13ea5