[roll] Roll fuchsia [assembly] call product assembly during the build
This allows the product assembly code to be exercised, and allows
the migration of features from assemble_system() to product
assembly, for use out of tree.

The image assembly configuration generated by product assembly
is validated against that created "the old way".  This is a
check to ensure that an equivalent configuration is being created.

Original-Bug: 91743
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/634546
Original-Revision: 31213ca1d4c39f4a1651c7ad6987f2626cdd4cd5
GitOrigin-RevId: 86ac7250adb8f2c5fb3c929c083481497b6f04a4
Change-Id: I3973b6ca18031084e9e766db5e656c18215aa29c