[roll] Roll fuchsia [inspect][codelab] Make part 1 and 2 of codelab more consistent
The codelab had some inconsistencies that were introduced with the
migration to CFv2. This change removes the inconsistencies in part 1 and
2 and fixes some typos.

Additionally since CFv2 the FizzBuzz component for part_2 of the codelab
needs to be added via the component Binder otherwise it's not started.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636501
Original-Revision: dc802b885cdb059cd1cdde0590bee7549e2a6b5c
GitOrigin-RevId: 5aaea7d80351dfc1e926209e19cf940714637c42
Change-Id: I2afc28b9734a5f332d34a6e65afa0bbf0edf400a