[roll] Roll fuchsia [ZBI Parser] Slightly tweak the get item API.
The item services use the extra field as a key, so we need to include
the extra field in the ZBI parser result.

This is also generally more extensible, as we could now trivially add
more metadata to the result as needed.

Original-Bug: 44784

Tested:
Unit test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639926
Original-Revision: f6c609a9ba5af9303c5723146629d08572ec37f8
GitOrigin-RevId: bf7944170d1277dc6140f2a16a24e42bebdbf747
Change-Id: Ifa2389fb10a3b76bd4078f3e7157bcad3b84e3d7