[roll] Roll fuchsia [fdf][compat] Prevent UAF in callback
We cannot pass raw pointers to FIDL callbacks, as we don't
have a guarantee that the pointer will be alive when the callback
is called.

Instead we should print the error in a future, as the future will
be cancelled if the class goes out of scope.
We have this guarantee because we wrap our futures in a 'scope' object.

Original-Bug: 91222
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640119
Original-Revision: 5fee3acaecfce31f56d85d532b0ce7e9fd9c571a
GitOrigin-RevId: c9518d5ecb79ed53546b022958f005c5c029d65d
Change-Id: I04adbd237184939136a9fc630422d4f200dbae0d