[roll] Roll infra/recipes [checkout] Tweak description of the cherry_pick_patches flag
Change the description of "newer commits which are unlikely to be
compatible with the older revision" which isn't accurate.

Refer to "the commit from the Gerrit CL" rather than "patches" to be
more specific.  Also use the singular rather than the plural to be
more specific about what is going on -- a single build only applies
one of these.

Original-Bug: 87199
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/638572
Original-Revision: f6234ca9bcc360d8a71907591080d1e598b535a8
GitOrigin-RevId: 42fa3e6d2b8c10e3986d9b69dea688059c1d2ebf
Change-Id: I05217b35f7476807cdc388622b8914ee2f80212b