[roll] Roll experiences [ermine] Move Ermine to the account storage capability.
The long term plan was always that components inside a session should
persist their data on the "account" storage capability. These
components potentially deal with sensitive user data so it is
important to encrypt these data at rest using the user's password,
which is what the "account" storage capability does.

Bug 92479 in CF actually means the "data" storage capability isn't
currently working inside a session, so moving over is more urgent.

Note this also adds a single startup log lines for ermine to report if
it found settings or not. Happy to remove this if needed.

Original-Bug: 92479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/639765
Original-Revision: 004ba47ab71628ffff46c2996535259443457091
GitOrigin-RevId: 44004e39042ce906bf0a461c04a015b54dbcb1c3
Change-Id: Ica8465411713ceca9816436b2240743ff6ab7c1c