[roll] Roll fuchsia [ftl] Add a retry and some reporting
There is evidence that suggests we have prematurely stopped processing
the last map block due to an intermittent failure. To hopefully catch
this and the reason for failing, we will retry reading the last page and
see if we succeed on the second attempt. If we do, then we will report
that we succeeded (via Cobalt).

There is no easy way of testing this automatically, so I made a simple
tweak to simulate a failure and then verified that the metric turned up
via iquery. This won't appear in Cobalt until we update Lapis
configuration. This change is likely only temporary.

Original-Bug: 87653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639924
Original-Revision: bf89b55d3d127f857f33570d0ef45be7c742b2a5
GitOrigin-RevId: efd4ec5d621bc9f6be044ea5019e08f9aa3676d0
Change-Id: I32e24fb1e6fde6099e4109ae6247255e8f2bb73e