[roll] Roll fuchsia [fdf][compat] Check if Node is valid after Bind
A node can be freed while we are waiting to see if it can be bound.
This can happen if the node's parent has decided to exit.

In fact, this has a chance at happening in any DFv1 compat test,
because the last driver of the stack will produce a node that
checks if it can be bound. It's possible for the test to finish
and start exiting before that last node is checked to be bound.

Multiply: compat-driver-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640202
Original-Revision: a1e47164591cd7d9800ecafc3e9a039d4bd92c6d
GitOrigin-RevId: 2b631735a8e1b8fd41e0172a8da9697d783873dc
Change-Id: Ic0c9f4e71e425fabffab61a62fa18f73dd335e29