[roll] Roll fuchsia [reachability] Rework "Gateway" criteria
Consult the route table to collect routes that have a "next hop" and
treat those addresses as gateway addresses. This is in contrast to the
previous logic that incorrectly considered assigned addresses' subnet
mask in attempting to determine if a suitable local address can be
selected for the route; this is the responsibility of the network stack,
so can be delegated to it.

Ping gateways concurrently while I'm here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640142
Original-Revision: 9c65d60e465da5962fa46e7b0c281ac5f367b3ca
GitOrigin-RevId: 31ac4582c95c0097adc001c6667061fd4e25778c
Change-Id: Ia277c6a54d2ef1363db495b764752990beb1af57