[roll] Roll fuchsia [mold] Abstract buffers.
This patch updates the way buffers are used in mold. It aims to abstract over
the possible layouts that the Buffer type can have. For the moment, only
LinearLayout is implemented but it will allow us to implement other layout
types in the future.

Every layout type has a TileWriter associated with it that will be in
charge of writing in the buffer.

When rendering, we iterate over each tile and create an associated
TileWriter instance depending on the layout.

This patch introduces:
- a new type, SplitsCache, that encapsulates cached buffer pointer usage
  in a safe wrapper
- a Layout trait to create new layout types (e.g LinearLayout)
- a TileWriter trait used by the Layout to write to the buffer (e.g
LinearTileWriter)

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/613647
Original-Revision: 7edfc405dd0071b790d6f55727123b9b65cc0a67
GitOrigin-RevId: 3e78a15b8443dfca976aac67eef5fa291941a919
Change-Id: I2d8c0b475fe4bddef6874dc0d2bebcfd48db0e62