[roll] Roll fuchsia [cm_moniker] Rename ExtendedMoniker to InstancedExtendedMoniker
This CL renames ExtendedMoniker to InstancedExtendedMoniker to signify
that it contains a ComponentInstance variant with an
InstancedAbsoluteMoniker. This continues to live in the cm_moniker
library; it is used by component_manager and allowed dependencies that
require an instanced absolute moniker from an extended moniker.

The type ExtendedMoniker is available in the moniker library; this type
includes a ComponentInstance variant that contains an AbsoluteMoniker
without instance ids. This is the preferred type to use for development.

This CL is a rename only.

The reason for this CL is to isolate the areas that require an
InstancedExtendedMoniker (primarily consumers of the event system), from
areas that do not need an instanced moniker variant from extended
monikers (to be addressed in a future CL).

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639988
Original-Revision: 80ddbd06d265983b7531181a1e22f683406be0a0
GitOrigin-RevId: b0f48dd82422bf740fd18c70f76b57c4f6e12cf9
Change-Id: Iae42da9a9ce1f422597f8f4d620d7b3c620459fd