[roll] Roll fuchsia [ffx] Update argh documentation for ffx emu
As we move toward releasing the `ffx emu` tool to general use, we need
to update documentation to make it easier for the new users to know what
to do. This CL starts this process by adding details to the help-text
for subcommands and their args, which is also propagated to fuchsia.dev
and will serve as the starting point for other doc changes moving
forward.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640122
Original-Revision: f029871477b2b1cb7940d92d3703973599e461cc
GitOrigin-RevId: 08236d5d5ffb8511197139ec5174505330106076
Change-Id: I9788e2d8f21fec6553bc92fb67960bdb5a68ac01