[roll] Roll fuchsia [netstack3][pmtu] Split contexts
This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See https://fxbug.dev/92605 for more details.

Original-Bug: 92605

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639789
Original-Revision: dfc42b86aa3667c1b87bbe900baddf5c883b576f
GitOrigin-RevId: 5722e65eaaabce06bbd5000074795149e648c5dc
Change-Id: I6fcdcbdad0cf769ee2ea7102abdf5f93a5ee81a9