[roll] Roll fuchsia [audio_core] Allow configuring a default SW gain for inputs and outputs
This CL makes it possible to configure a default software gain for input
and output devices. This will be used on some products to boost the
microphone gain.

Some background on how gains are set currently:

- The config file has a driver_gain_db field which is always applied at
  the driver level.

- The AudioDeviceEnumerator.SetDeviceGain FIDL call currently updates a
  SW gain knob. It does NOT update the driver gain, which can be set
  only by the driver_gain_db field mentioned above. However, the
  SetDeviceGain request is filtered through what the HW allows -- the
  idea is that we might change SetDeviceGain to update HW gain in the
  future, so we should respect the same limits.

To fix the linked bug, we need to apply a SW gain because the explicit
goal is to apply gains that cannot be applied in HW. Hence, this CL uses
the same code as SetDeviceGain except that it ignores the HW gain limits
check.

This CL may lead to a potentially-confusing scenario: if a client calls
SetDeviceGain, this will override the default SW gain but not the
default HW gain. This potential confusion cannot be avoided without a
update to our device APIs, which is in progress separately. In the
meantime, we don't expect this to be a problem because the only
SetDeviceGain caller is the vol tool, which is for developer use only.

Original-Fixed: 90496
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/621981
Original-Revision: b0078b133d2fc74f1ed92e7a0a27a1036cb6c6b8
GitOrigin-RevId: d7edba1810b710598e5e4a8a4ca330a463325375
Change-Id: I884ebb9f81cf8f9d272288df9ee9576db38d6cad