[roll] Roll fuchsia [component_manager][component_tree_stats] Stats keys on non-instanced monikers
This CL has diagnostics stats use monikers without instance ids to key
on for component instances.  `InstancedAbsoluteMoniker` and
`InstancedExtendedMoniker` are replaced with `AbsoluteMoniker` and
`ExtendedMoniker` respectively.

When the Started event triggers the `on_component_started` hook, the
hook will not return early if the stats tree already contains the
moniker, and a `diagnostics_waiter_task` will get spawned for the
(duplicate) moniker. This results in a task measurement created for two
dynamic instances that share the same `AbsoluteMoniker` to share the
same key in the component stats tree.

This change depends on a Started event having a 1:1 mapping with a start
action (ie there cannot be multiple started events dispatched from one
start action).

The reason for this change is to reduce the use of monikers that include
instance ids in business logic where possible.

Original-Bug: 73171
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640154
Original-Revision: 2795196d5d8ef6b2c65a19c88eeb608c04108163
GitOrigin-RevId: 3a7c48bef6671e5583af25065d451d906dbaadb4
Change-Id: Ibd115a4ddb1e41ead0968b8605aee000409d767d