[roll] Roll fuchsia [fxfs] Fix journal corruption
If journal replay advances the file offset to a non-block-aligned
offset, the writer was erroneously configured to start writing entries
immediately at that offset, rather than the next block-aligned offset.

This is in incorrect, because then the journal writer has to
read-modify-write existing journal blocks, which can result in
corruption since the block writes might not be atomic and existing
serialized records might end up in an indeterminate state.

Fixed this so that the journal writer always skips to the next
block-aligned offset.  This also has the benefit of preventing the
journal writer from doing costly read-modify-writes.

Re-enables `fxfs-crash`, which reproduces this issue.

Original-Bug: 91684
Original-Fixed: 92435
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641041
Original-Revision: 9fac8be1639fcee5d463d505e8d53ae29f2493da
GitOrigin-RevId: 2c7cea670ed2e5ace09fbb9a2936e07543279b17
Change-Id: Ie9caaed27b40c689a8509445517c909fa323e92c