[roll] Roll fuchsia [bt][pairing] Signal on responder recalling key
When we are the responder (the peer initiates the pairing request) and
we have a key to provide, the pairing state previously never signaled
completion of pairing, meaning that the "initiating connection" never
was cleared out.

This resulted in the upper-layer state never being "connected" after
the semantic change to require a pairing success.

When we are waiting for encryption to be enabled, it is possible our key
is not the correct key. Allow the peer to re-start the pairing process
in this state.

Original-Bug: b/213331342
Test: added unit tests; fx test bt-host-gap-tests
Test: confirm by connecting from peer and observing "connected" state
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/631023
Original-Revision: da6d2692024bf804c797bb70e0b61d3eebb74c78
GitOrigin-RevId: e0152f4a81fb358b5bf2cf4841d61d9ac75fc2f5
Change-Id: I208fc0805ee2e4dc0f698e12a2383873e494a834