[roll] Roll experiences [login] Fix data persistence race when launching ermine
This change fixes the issue when ermine is launched before empty
password authenication is done. The error was in the conditional logic
used to determine when login is done. It was triggering ermine launch
when just oobe was bypassed (a hold over from pre-UDP days).

Original-Fixed: 92629
Tests: Manual test with both OOBE = OFF and ON
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/641181
Original-Revision: 400faf3956f56b6ab41fa11f631673ff3b10f933
GitOrigin-RevId: b3e97dd60022bb50e96d60f1822e2bb244eedb79
Change-Id: I76dbbd6480d42e36bf187c3860942df94ee1aad9