[roll] Roll fuchsia [netstack3][reassembly] Split contexts
This is part of a larger change in which all context arguments in the
Core are being split in two in order to resolve some lifetime conflict
issues and to prepare for a deeper overhaul of the way state is accessed
and synchronized. See https://fxbug.dev/92605 for more details.

Original-Bug: 92605
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639790
Original-Revision: dbc0541b47e8c37964488af679983af54ffcb3c9
GitOrigin-RevId: b79d64a1eb053b9171728f9da9fa8f5a58d4519e
Change-Id: Ife13fe91aec6db4bc10e59aa49b94042f8973361