[roll] Roll fuchsia [archivist] Distinction between external and internal event producers
This change introduces two types of event producers:
1. Internal: emit events that are purely internal, they may or may not
   originate from external actions.
2. External: emit events that come from external sources (such as v2
   event source, v1 component event provider, unattributed log sink
   connections, etc).

The goal of this is to provide a mechanism for shutdown. During shutdown
we perform the following actions before ensuring all logs are drained:
1. Stop external event ingestion
2. Drain all external events that were already buffered.

This change provides the the API to accomplish that, the following
change will exercise it in the archivist.

Original-Bug: 92016
Multiply: archivist-unittests, archivist-integration-tests-v2

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640123
Original-Revision: 5f6e5891e04f000b3277a089c7e050a28370583c
GitOrigin-RevId: 922c0912e7fb3fbe0af8c5d53b68176faf097f78
Change-Id: I8b6c0e036c60deea0c4c18161a57b1416a665765