[roll] Roll fuchsia [archivist] Use new event router
This is a first step towards the event bus architecture for the
archivist. In this CL we refactor the existing event ingestion to use
the new API.

A lot of the advantges of the new API aren't still clear since we still
don't decouple repositories and have them emit events too. The event
router producer and consumer installation happens in diverse places in
archivist.rs, the goal is to bring them to the same spot so that it's
very clear how the connections are made.

Multiply: archivist-unittests, archivist-integrationt-tests-v2
Original-Fixed: 92016

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/635924
Original-Revision: db4d52cadff50fa815d9b6e95e07ae0f517ba4fc
GitOrigin-RevId: 02e7b75d894bfabee20a542763a6e89c24258811
Change-Id: I944d0f6c007d8ee43d984156e6ff2531bb97b17d