[roll] Roll fuchsia [llcpp] Support encoding to multiple wire formats
Currently LLCPP does not provide a way to specify which wire format to
encode into. The wire format is determined simply by whether the message
is transactional or not. This CL adds support for specifying the wire
format version used in encoding.

The motivation for doing this is to make it possible to migrate
persistent encoding over to v2 piecemeal.

Work on FIDL at rest will introduce new APIs for persistent FIDL.
As a result, these APIs will soon be marked unstable and should
not be used out of tree.

Original-Bug: 82681
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639999
Original-Revision: dfe9505a3e18f85b2944d878fae39832cfa1bf71
GitOrigin-RevId: 220e39b8ce0fe7a40dfb114055c76f004df3c904
Change-Id: I7de8417e9fcad93a80692b294218961811bbf877