[roll] Roll fuchsia [netstack3][reassembly] Make tests more flexible
Make the IP packet fragment reassembly tests more flexible so that we
can test receiving packets which have nonstandard values. This allows us
to write a test which mimics a sequence of received packets which were
found through fuzzing to cause a panic. This commit adds this test, but
marks it as `should_panic`. A future commit will be needed to debug the
issue, fix it, and remove the `should_panic` annotation.

Original-Bug: 92831
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641021
Original-Revision: 256caec2ed1eac155224f42fc22156a9e092111e
GitOrigin-RevId: 8fedc762166ce44e60fb2d5cd6ebb488b597df8b
Change-Id: I94723f7fb351129335d00ac0c72063fe88a99dfb