[roll] Roll fuchsia [zxcrypt] Issue BLOCK_OP_FLUSH on Shred
Shreds really do need to make it out to the underlying storage, and we
can't rely on anything else ensuring that our writes are made durable,
so we should request an explicit flush before returning.  It also makes
sense to ask for one as we format a volume, to ensure that both birth
and death of a volume are durable.

Original-Bug: 92790
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640768
Original-Revision: 70cc05ba91283abc3678ded6e4ca1c36e3ad96cf
GitOrigin-RevId: 94acc1f3611609a4d7d54fcecb98c7353bb472f7
Change-Id: I166b64215755f186bd93d4daa4ac12de551f617a