[roll] Roll fuchsia [netemul] Do not log ERROR on channel closure by client
Currently, if a client of the netemul sandbox closes their end of the
fuchsia.netemul/ManagedRealm protocol with an in-flight request, the
sandbox logs an ERROR and propagates the PEER_CLOSED error upwards. This
can cause test flakes and is unnecessarily severe.

Instead, log the PEER_CLOSED error at WARN and do not propagate it.

Test: fx test netemul-v2-tests
      fx test netstack-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641621
Original-Revision: 7ae14e6987e0af5c0605756be74dece4d20925cc
GitOrigin-RevId: 9780fa324285d45755a520d0eee771a0d0860c39
Change-Id: I34849b077cc83b8b1323c1786d5bd1b8220ddafd