[roll] Roll fuchsia [fxfs] Move the crypt service into stores
The crypt service is supposed to be attached to the store rather than
the filesystem.  This CL makes that move.

This CL also changes the API for create_key: instead of supplying a
wrapping_key_id, the key service should return it.

Original-Bug: 92275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640402
Original-Revision: 571b20cb311b8cb666d7474c213d9774fd19d708
GitOrigin-RevId: 923ed5a96b1dd9523ec993a88e343a00953eb2ad
Change-Id: Ic1402d07e60ac2e0d19b4bbf8eaf3bf77dd10961