[roll] Roll fuchsia [netstack3] Use a single concrete IpSock type
Previously, we allowed testing code to use a different type for IP
sockets. This genericity turned out to be unnecessary, and caused
problems when trying to contextify the IP layer. In this commit, we
hard-code the `IpSock` type everywhere, and get rid of the testing-only
`DummyIpSock` type.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585921
Original-Revision: 8001c0cf44a3108de40495ff67459398ceb1876d
GitOrigin-RevId: 505be7ead18f37e8d726ebfe9ba1e3f1572825f0
Change-Id: Iafbb07ac85173403cfe0c743eceef04ac54bab00