[roll] Roll fuchsia [network/tests] Fix flake
...in add_remove_address_on_loopback.

See comments inline.

On Netstack2, the DAD resolution event (which the interface watcher
depends on to update its view of interface properties) is handled
asynchronously w.r.t. the add address operation so the new IPv6 address
may not be observed in the initial "Existing" event; it may be observed
in a "Changed" event.

This is not an issue on Netstack3 as we will observe the address in an
"Existing" event right away as we do not perform DAD and Netstack3 is
queried for all interface states on watcher creation (no events happen
asynchronously).

Multiply: netstack-fidl-integration-test: 100

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641941
Original-Revision: 643ca87abb9500cf7fe9794fe22a5c494ce60238
GitOrigin-RevId: d94a3c623f30c9e2aa239c888e2689d27bb1680f
Change-Id: I3c50e27e57b5e97e4d6b4584be7b9ec609b03886