[roll] Roll fuchsia [outgoing_directory] Introduce prototype library
This CL introduces a new version of the `service::OutgoingDirectory`
with certain differences. First, this class will live in the
`component_llcpp` namespace. Second, it will use `//sdk/lib/svc`
as a backend instead of the VFS static library. By swapping out
the backend from a static C++ library to a C shared library,
the space costs incurred by using this library will be minimized.

This library will be iterated upon until it's at parity with the
`service::OutgoingDirectory` library. Once we've reached that point,
clients of that library will be migrated to this one, and we'll
deprecate and remove `service::OutgoingDirectory`.

Original-Bug: 89036
Original-Bug: 92898

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/636566
Original-Revision: d92e83262707643509ad2992de21de25e58cd850
GitOrigin-RevId: db0ec2d87f45dc4cbe4daeb880f3f79c7abb6896
Change-Id: I15cea74e83c0e522183265ad7b76d56531da2829