[roll] Roll fuchsia [virtualization] Provides tools to wait for guest pty
It is not sufficient to wait for the guest console to Start before
interracting with a guest over serial as, for example with a Linux
guest, the pty might not be started and so input might go to a
black hole. The EnclosedGuest framework solved this internally by
having repeatedly echoing until the pty was definitely up.

This moves a version of the EnclosedGuest helper into the
GuestConsole so it can be used by other test frameworks, and clarifies
that comments on the Start method to indicate its limitations.

Original-Bug:86513

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642041
Original-Revision: ab08de8f592e8d25ef95011f70e0f48ade8b7098
GitOrigin-RevId: df08c0778636fe4f9221959ac4d827dd367bb0b8
Change-Id: If4bf100887c12f8f35050cf3ef581057af0c90da