[roll] Roll fuchsia [kernel][vm] Retain backlink when pinning
When a pager-backed page is moved to the Wired queue on being pinned,
retain the object backlink and offset. This allows us to validate the
object backlink irrespective of the queue the page is in, as long as it
is owned by the VmCowPages object. The vm_page_t's object field can be
shared by a stack owner if it is loaned, so this enables us to check the
backlink when that is not true and the page is owned by a VmCowPages.

Test: KernelUnittests

Original-Bug: 92905
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641396
Original-Revision: 32e68a6853bffca4bbd6a064e99bf2a5dd0208fd
GitOrigin-RevId: a4f8b35c48e3b9e4347016111f1192b319c9739b
Change-Id: Ie6a809d37d7967266cead6e563cbe841e5785da6