[roll] Roll fuchsia [realm_builder] Split up shard to base/internal
Since usage of the Realm Builder client libraries relies on adding a
child, Realm Builder Server, with a relative URL, the shard is split it out
into two parts. This is necessary because OOT environments won't be able to
include the Realm Builder Server in their test package.

In order to remedy this problem, the shard is split into a base shard
and an internal one. The base portion contains all the shared routing and
capability declaration that is used by the test component's environment.
The internal one inherits from the base shard, and extends it by declaring
the necessary child based off a relative URL.

This is viable in-tree because it can be enforced with internal build rules.
OOT, a similar shard will be necessary to allow for OOT test components to
use Realm Builder.

Original-Bug: 89252

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640568
Original-Revision: d1450fdb56fc1a0dcd2a5489fdfcc810534f167d
GitOrigin-RevId: 4de22fa52a4d58f028dfb8814c7917cc90a13492
Change-Id: I6dec8bd0ce7635cb32d00b292863d15471def95c