[roll] Roll fuchsia [rust] Remove unsupported target features
Apparently rustc doesn't recognize armv8-a or simd /shrug

We get a lot of this stdout spam:
```
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
[18801/31788] RUST obj/garnet/lib/rust/io_util/libio_util.rlib
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
'+armv8-a' is not a recognized feature for this target (ignoring feature)
'+simd' is not a recognized feature for this target (ignoring feature)
```

Originally I thought this is necessary to get Rust auto vectorization to work.
In hindsight I'm not sure it's necessary to specify +simd.
The godbolt links here just say aarch64 which I guess is already implied?
https://www.nickwilcox.com/blog/autovec

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642001
Original-Revision: b43c34cb5e2b41e608f93add7ffe02c7ba8f337c
GitOrigin-RevId: 3373aedd42dfa77ccbbf90fc62927cd4d25361d5
Change-Id: Ib192c9a0605a9187bfa5d1c19963623d7265da65