[roll] Roll fuchsia [audio][drivers] Add indirection to access signal processing APIs
Instead of directly composing the SignalProcessing protocol, we
add a method to acquire the protocol. With this the signal
processing APIs can evolve independently of servers of the
StreamConfig, Dai and Codec procotols which do not support
the signal processing APIs. These servers only needs to implement
the mechanism to get the signal processing APIs instead of
reporting that they don't support each method in the signal
processing protocols.

Original-Bug: 64877
Test: simple-codec-unittest
Test: tas58xx-test
Test: nelson-brownout-protection-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640104
Original-Revision: a2608908869e8ff21c7628822e113a36ea649e60
GitOrigin-RevId: 5f91487d64110d509a10777cdf58339a1bdb48cc
Change-Id: I3730f63a738fdec5ac28246d75a6af2d056ba28b