[roll] Roll fuchsia [component_hub] Timeout on `runtime` dir exploration
The `exec/runtime` directory is assumed to open or return
PEER_CLOSED. There is a third option: the directory hasn't
been served by the runner. In that case, the library will
hang waiting for a response.

To fix this, add a timeout to `runtime` dir exploration

Original-Fixed: 93129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642207
Original-Revision: 02f6527909d522f6cf7dbca5115a3d528bbf2572
GitOrigin-RevId: 72efcc30210e1761d7518d7669c5817cd7a2fdb4
Change-Id: I00aafca18bf35b98a5d9beb0cf013b715de6520a