[roll] Roll fuchsia [fidl][llcpp] Change SDK back to experimental
The category of the SDK was unintentionally set to partner in this
CL - fxrev.dev/609051. This sets it back to its original state.

Note - with category 'experimental' the ddk sdk would not compile
because of level mismatch. Until this is resolved, make a local
change to set category to partner.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643042
Original-Revision: b71e6ea2ef89d0a5b2d2c2f8166aa58cbe0f5364
GitOrigin-RevId: c80c445ca399f0b81c8b56ef8b6ef32b8785ee7a
Change-Id: I44776856d9408d4c59b2e69cd9ff3a55febafc8a