[roll] Roll fuchsia [iwlwifi] remove the duplicate command
The first iwl_mvm_configure_filter() call is no-op because
the mvmvif->bss_conf.assoc is still false (checked in
iwl_mvm_mc_iface_iterator()).

This is the code left unexpectedly during a refactoring.
Remove it in order to remove the error message in the log.

Tested: The unittest code is unchanged because the first call is no-op.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642233
Original-Revision: d2b87487525ec7815d1fcb86d47db487c2598539
GitOrigin-RevId: d6c5d489276e687c6edaf970cb8dcb91dc28f2c1
Change-Id: Ic8c4d0d05505e39679c24483cf2fdcedafc8bb82