[roll] Roll fuchsia [basemgr] Add support for starting v2 children
After this CL, basemgr will be able to start all of its
children using the `fuchsia.component.Binder` protocol.
Furthermore, it will listen to the channel connection
of this protocol to observe for component crashes. It
will attempt to restart the component (by reconnecting
to the protocol) up to 3 times. If it fails to start
after the 3rd attempt, basemgr will trigger a reboot.

basemgr will only start children that it is configured
to do. In order to instruct basemgr what children
to start, the flag `--child` must be used. basemgr
will use this name to determine the path of the Binder
protocol in its namespace. A sample manifest of basemgr
with children 'foo' and 'bar' would look like:

```json5
{
    program: {
        args: [
            "--child=foo",
            "--child=bar",
        ],
    },
    children: [
        {
            name: "foo",
            url: "...",
        },
        {
            name: "bar",
            url: "...",
        },
    ],
    use: [
        {
            protocol: "fuchsia.component.Binder",
            from: "#foo", // Assuming foo is the child name
            as: "/svc/fuchsia.component.Binder.foo",
        },
        {
            protocol: "fuchsia.component.Binder.bar",
            from: "#bar",
            as: "/svc/fuchsia.component.Binder.bar",
        },
    ],
}

```

Original-Fixed: 72349
Original-Fixed: 90960

Mutiply: basemgr-v2-test.cm

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639025
Original-Revision: eca3133a6912d3fac683c48484d46981247d25ec
GitOrigin-RevId: ba6286baecc3e584e59320f61c6609dbf4257f06
Change-Id: I2336f6dd317cbc94fecf9c679fbfcc29090d8e67