[roll] Roll fuchsia [async-loop] Use worker threads to call cancelled closures
When an `async_loop_t` is being shut down, it needs to cancel all
pending tasks with a ZX_ERR_CANCELLED status code. The async loop
currently calls closures from the thread performing the
`async_loop_shutdown` call (or, in the C++ world, running the
destructor).

This existing behaviour is problematic for thread-hostile code: on
a single-threaded dispatcher, code using the dispatcher may reasonably
expect that its closures are _always_ called on that single thread,
independent of whether the callback is being called as part of normal
operation or if it is being called as a part of a shutdown operation.

This CL adjusts the behaviour of async loop shutdown to cancel all
outstanding closures using one of the loop's worker threads. Only in the
case that no worker threads are present (e.g., no thread was ever
attached, or threads have already exited due to a previous `quit`
operation) will the callbacks be called from the shutdown thread.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/611889
Original-Revision: f591b77659d3c21d99bf1289d1e030c818cae299
GitOrigin-RevId: 4d694616ddb165cf4cee5bcad13022ace51477d2
Change-Id: I8832e509448a90a577c05936884ef587f1f55e2d