[roll] Roll fuchsia [routing] Allowlist detection does not use InstancedExtendedMoniker
The source_moniker stored in `CapabilityAllowlistKey` now refers to an
ExtendedMoniker that does not contain instance ids in the moniker for
the ComponentInstance variant.

`verify_routes_exceptions_allowlist.json5` is modified so that the
`component_instance.path` does not contain an `instance` field to
correspond with the non-instanced moniker.

The reason for this change is to reduce the unnecessary uses of
instanced monikers in the public API space.

Original-Bug: 87407
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639231
Original-Revision: 4da743153b557a984a0bf9950f7dc9da8fb40710
GitOrigin-RevId: e9c84c3915f949ef04b2293149582026b3e14277
Change-Id: I5756613688359f3b7e12777c4a8904285455a28a