[roll] Roll llvm-project/libc [libc] Replace type punning with bit_cast
Although type punning is defined for union in C, it is UB in C++.
This patch introduces a bit_cast function to convert between types in a safe way.

This is necessary to get llvm-libc compile with GCC.
This patch is extracted from D119002.

Differential Revision: https://reviews.llvm.org/D119145

GitOrigin-RevId: 754a17d590a0c7354391f918dde4ebe9d68cf9c2
Original-Revision: 6e3e44fa21091d88e69fbaf51b62d42b65fefdc3
Change-Id: I182e384c3a76415cf2f6586e328ff2110a3f9388