[roll] Roll fuchsia [kernel][sched] Redefine WaitQueue's total ordering.
Redefine the total ordering of a WaitQueue to no longer be static and
based on priority, but instead to be dynamic and based on a
combination of |now|, and the specific details of the scheduling
discipline and scheduling state of the threads blocked in the queue.

See the (lengthy) comment added just before WaitQueueCollection in
kernel/include/thread.h for details, but the basic idea is that:

1) The most important thread to wake from a wait queue is the deadline
   thread with an absolute deadline in the future, closest to now.
2) If there are no threads with absolute deadlines in the future,
   choose the deadline thread in the set of waiters with the minimum
   relative deadline.  This is the thread which _would_ have the
   minimum absolute deadline were all threads to be woken now.
3) If there are no deadline threads at all in the queue, choose the
   fair thread with the minimum virtual finish time.  Assuming that
   all fair threads currently blocked in the queue are eligible to run
   right now, this is the same decision the scheduler would have made
   when evaluating eligible threads in its run queue, and helps to
   ensure that even if a thread has high weight, it cannot be be
   picked before a low weight thread forever, even though it does get
   a boost because of its weight.

Also note that it is possible in this system for step #2 to choose a
deadline thread which is not currently eligible to run.  This is
because the scheduler does not currently demand that the absolute
deadline of a thread be equal to when its period ends and its
timeslice is eligible for refresh.

While it is possible to account for this behavior as well, doing so is
not without cost (both in WaitQueue object size and code complexity).
This behavior is no different from the previous
priority-based-ordering's behavior, where ineligible deadline threads
could also be chosen.  The ability to specify a period different from
a relative deadline is currently rarely used in the system, and we are
moving in a direction of removing it entirely.  If the concept needs
to be re-introduced at a later date, this data structure could be
adjusted later on to order threads in phase 2 based on the earliest
absolute deadline the could possible have based on earliest time that
their period could be refreshed, and their relative deadline parameter.

Next up: completely remove the concept of priority from our threads,
PI, and scheduler behavior.  Use aggregate weight as a stand in before
introducing generalized deadline inheritance, and restrict the concept
of "priority" to the very edges of the kernel APIs, translating
priority into weight when creating profile objects.

Test: existing system-wide tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/615965
Original-Revision: 6f80fb1ab44b2b34d5ae4d6555cacac0e475ea8f
GitOrigin-RevId: 13b736a5c6e1887d1b23f55604e613fa3772aa2c
Change-Id: I90e3db63c1c04d6e4c172b7ffb7e43515e166c45