[roll] Roll fuchsia [tools][fuzz] Rewrite testcase paths in libFuzzer output
As a followup to fxrev.dev/642353, which rewrote incoming path arguments
from data/ to tmp/, this CL rewrites testcase paths back to data/ in the
libFuzzer output when running a testcase repro. This is not referenced
by ClusterFuzz during normal operation, but it is expected by one of the
integration tests.

This also fixes a bug in handling of libFuzzer args with dashes in their
right-hand side.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643286
Original-Revision: faf0d0761f334c54527551604ed4265fe148f927
GitOrigin-RevId: 1fad6084bcbccc1b53abdc1aeb1a3298de3eef09
Change-Id: I1e21642fd0eb5614fd6f3969f8bb71277f0b5d69