[roll] Roll infra/recipes [sdk] Use sdk_archives.json to determine files to upload
...instead of assuming that the targets passed to ninja explicitly list
all the archives, which will no longer be the case once SDK builders are
switched over to the high-level //sdk:archives bundle instead.

As a bonus, we'll no longer need to expose `built_targets` in the fint
build artifacts as this is the only place that field was used. This
will simplify the interface between fint and recipes and reduce recipes'
knowledge of Fuchsia's build system.

Original-Bug: 92108
Original-Fixed: 93299
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/644143
Original-Revision: 967a5748de564ad4f169209acef9cea5df47a880
GitOrigin-RevId: 851c615917b55e5da066e33e7acbd5a5fd622dcb
Change-Id: I38a8882e5994ed6b725d1f28134763fcc91dcb9c