[roll] Roll fuchsia [carnelian] rename and improve app context
AppContext was too generate a name for a struct whose sole
responsibility was sending messages to be handled in the executor main
loop. This change renamed it to AppSender and enhances it with a way to
use the AppSender from another thread.

Testing: button uses a thread to drive the resize messages.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642441
Original-Revision: 372c091c921e81d37649e40c869f6e2e3f77a5eb
GitOrigin-RevId: 55750342a09dc4981e49232f92ad4b7f237af5fb
Change-Id: I2c79420058b1210aae134f271c1fb403f43702fa