[roll] Roll fuchsia [docs] Fixing shell commands on the fx page
This CL updates the //development/build/fx.md doc shell blocks. Some of
the blocks had "$" characters before the commands the reader is meant to
copy, so when they pressed the block's copy button, the "$" was copied
as well. This meant when pasting the text, the user had to edit the
copied text.

In this change, I updated a few short command blocks to use
"posix-terminal" instead of "shell. This places a "$" before the text,
but it doesn't get copied to the user's clipboard, which makes
copy/paste more convenient. I also updated a few of the longer blocks,
which contained both code and comments, by removing the ability to copy
and paste.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643641
Original-Revision: 54c83ba3a46280466dde1542906d9b6180149eb8
GitOrigin-RevId: 92b39615475d2f85f11576adfa73f6162430451b
Change-Id: I9b272923a2ef3c4fcf9ed25eb1f7de748e965543