[roll] Roll fuchsia [kernel][lib/arch] Conditionalize hwreg::EnablePrinter use
hwreg::EnablePrinter is a nice convenience that comes with a large
runtime overhead in object size.  It's mostly used only for
build-time generator programs, and elsewhere only in test code.
So use a build-time conditional in lib/arch headers to enable
the printing support only in the places that need it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644041
Original-Revision: fe47fb7cac450a7f06eea084e6f19eeb546a287f
GitOrigin-RevId: 41ab5b1b33b24dec3d69751d0da425b53b7894e9
Change-Id: I73520377204b93d4516352c174207b1d18887bb6