[roll] Roll fuchsia [ftl] Don't timeout after 60 seconds
Timing out after 60 seconds is problematic at the moment because it's
not possible to be certain that the callback will never be called or
that the operation has been cancelled.  The first will almost certainly
result in memory corruption and the latter has the potential to cause
confusion to higher layers.

Original-Bug: 87653
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643903
Original-Revision: 68157c9c71fbf10fe063bb15ac07cc91a3985fa9
GitOrigin-RevId: ef15de15be5524d056d93976b9cfa96b1ae3c29f
Change-Id: I6d0047889454033442ac1c0b909c85a0ac38f15f