[roll] Roll fuchsia [build][rust] assert rust edition is 2015 or 2018
We don't yet support rust edition 2021 in tree yet, but we didn't actually
assert that the edition value was 2015 or 2018. This allowed targets to
specify 2021, which would just fall back to 2018. This adds an assertion,
and switches all the targets that set `edition = 2021` to `edition = 2018`.

Original-Bug: 90449
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644263
Original-Revision: 493098e0dd162cc945f2dda5a34cc17d649657f6
GitOrigin-RevId: 997d3eb23fd696ef2376d9c339fc3a472eb0c8a4
Change-Id: I719b6f3acfa54c542f84ec6497eb951f46d714d7