[roll] Roll fuchsia [forensics] Prevent exceptions from going into oblivion
OnException is a "move-only" method and any exception sent via the
method can't be reclaimed if there's an error with the server. The
exception handling protocol has been amended to include a method that
allows clients to check if the server is able to respond to requests
before they make the call to OnException and cede ownership of the
exception.

Note: A response to IsActive doesn't mean an error didn't occur between
calls to IsActive and HandleException, but it dramatically shrinks the
window during which errors can occur and not be recoverable.

Original-Fixed: 57111
Original-Bug: 93282

TESTED=`/boot/test/crashsvc-test`
TESED=added crash to exceptions.cm and exception was released.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643786
Original-Revision: c0d7e9499360c57b9d7d2700f283b4447547a1b2
GitOrigin-RevId: 56eff6c72e2a970063fa9cd1d033835c06dbda13
Change-Id: I5728ca0295237688bec85fc918eb12c104d62b88