[roll] Roll fuchsia [terminal] Improved scroll bar
Switch to an overlay scroll bar that is hidden when
not needed. This means that the scroll bar is no
longer reducing the size of the terminal and it
provides a more immersive experience for apps
that use app-cursor mode.

The scroll bar thumb is also using alpha so text below
it is still visible. Thumb alpha is adjusted depending
on the location of the mouse and if the thumb is
grabbed for scrolling.

This also makes the terminal handle mouse and touch
input better in order to enable selections using the
mouse in the future.

Test: terminal_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644531
Original-Revision: d1c56c91ebc21b17ca5feca98749747add9ebe80
GitOrigin-RevId: 8b61922d4126f27ed2999ca2f2cc1825adcf4b60
Change-Id: I07c0afc98dcf64680c6aa3cf6c9258edcfdd3540