[roll] Roll fuchsia [wlan][softmac] Move WlanInfoBandInfo to WlanSoftmacBandCapability
The WlanInfoBandInfo struct is only used for SoftMAC, and so it's
moved to fuchsia.hardware.wlan.softmac and renamed to
WlanSoftmacBandCapability so that the arrangement is similar to how
fuchsia.hardware.wlan.fullmac defines and uses
WlanFullmacBandCapabilities.

Additionally, this change renames the bands and bands_count fields to
band_caps_list and band_caps_count.

Original-Bug: 92608
Test: None. Refactor only.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639961
Original-Revision: 3111db2957cd7657b3b0adcf1f21ea8125a60358
GitOrigin-RevId: da14ddac11186aeaab35d0eca79f2c731f5ca9e5
Change-Id: I1a2a6b88bf85a5ed836437321897b80dda83f92a