[roll] Roll fuchsia [bt][usb] Fix bug processing multiple CMD packets
When multiple command packets were sent to the transport at once, the
length of the buffer accepting the packets was overwritten and not
reset, resulting in rejection of a packet and channel shutdown.

Reset the buffer size before attempting to read from the channel, and
drop messages that are too large instead.

Test: added tests; fx test bt-transport-usb-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644421
Original-Revision: 675b1f0d7caf8b6c35aae3c6c5a4f5092c33876c
GitOrigin-RevId: 6038e538796f34d6a3c5288139e4331ff6ba7422
Change-Id: I45b5ff4fce1b5b9443afbe5428851076e37a0a8a