[roll] Roll infra/recipes [fint] Delete built_targets build artifacts field
It's no longer used by the recipes as of fxrev.dev/644143. Its only use
was to figure out which SDK archives were built by an SDK subbuild to
upload them to GCS, but we now use the sdk_archives.json build API file
to locate those archives instead.

Requires a bit of refactoring of the tests as they previously relied on
the build artifacts to determine which targets were built. Now we do
some hacky parsing of the ninja command line from the fake subprocess
runner to determine which targets were built. I also considered adding a
separate return value to the `buildImpl` to pass the list of targets
into tests, but with several return points in the function it's slightly
annoying to add a new return value. It also kind of nice to look at the
targets that were actually passed to ninja, rather than just assuming
the function builds the targets that it claims to build.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644161
GitOrigin-RevId: cd5893c540dcf5bcbe83448d235f5c922974b8a6
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/644354
Original-Revision: ed0c30df51d46b0c25383887fc88e15835cbe870
Change-Id: I89aa0f9baf322af112274e41d00700a6bdfd6b89