[roll] Roll cobalt Improve some of the Encoder documentation.
I updated the docs for the Encoder as it was a confusing mix of
Cobalt 1.0 and 1.1 information, and also to document the
EncodeStringHistogramObservation method. Also I fixed up some
uses of auto in the .cc file.

I also noticed the Observation type was missing documentation
for the new report type, so I added it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/644301
Original-Revision: 50c4ed67ef80a632900ff4f0a72876d429824ba8
GitOrigin-RevId: fe70fc6bee0a2063912fa39592dae417002299de
Change-Id: I089d133e645a159d5951d10578d57d788ae3972b