[roll] Roll fuchsia [docs] Fix RFC metadata
This CL fixes entries in _rfcs.yaml to use [''] rather than []. It also
fixes the rfcid in RFC-0088, and an incorrect heading level in RFC-0063.

The reason to prefer [''] is that the Jinja2 logic in _rfc_metadata.md
checks specifically for that. If [] is used, it emits an empty table row
instead of omitting it. While we could change that logic to check for []
instead, this default would likely lead people to edit [] to [12345]
(for example), whereas it should be ['12345'].

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644412
Original-Revision: 4d8370788cf144175698ec1705885f07b4fb2dd1
GitOrigin-RevId: 6e5912b7deeddbe1300bd98538c1c88b286e4aef
Change-Id: Iae09e9bcc7db29526016623577f17d486f853fe9