[roll] Roll fuchsia [paver] Add a FakePaver test util class
Move the fake paver implementation in netsvc into the paver library and
expose it as a test util class. This is to facilitate test of other
libraries that need to interact with the paver service.
//src/bringup/netvc and //src/firmware/fastboot will be two libraries
that use it.

Original-Bug: b/217597389
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644031
Original-Revision: a1cc3e9869d7eca35251beae8bb5cad2f8f1d6c7
GitOrigin-RevId: 36cdebd203c3a9c61cae8a6053ad1260e7950ff8
Change-Id: I4ba792f9e68cc7aeac781959a27e277d8ae401c7