[roll] Roll fuchsia [mdns] Add v2 component
This change introduces a CFv2 MDNS component under core, replacing the
existing legacy component that is in the sys realm under appmgr. It
leaves existing capability routing intact, so MDNS protocols are still
offered to the sys realm, and will be pruned if possible in a follow-up
change.

The legacy component remains because it is used in a suite of
integration tests that depend on the netemul runner, which has not yet
been migrated to CFv2.

Original-Bug: 90433
Test: fx test mdns-tests
      fx test mdns-netemul-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642888
Original-Revision: e2b912dd5bfe025021d7db120a509e8ab6266d34
GitOrigin-RevId: 1f22660d78f31dd7ad6e3b88daeb06239e859f1c
Change-Id: I79be04ffbb9586e08229726a041e0745b5215ef4