[roll] Roll fuchsia [clippy] Add all help messages to tricium comments
In addition to the main message of a lint, also add all the child help
messages to each Clippy Tricium comment. This will provide additional
information to CL authors (often including suggestions for fixing the
lint) and also automatically handles linking to the documentation for
the lint, as a link to the docs is often included in the help messages.

This also causes Tricium to stop adding broken links for undocumented
warnings like `unused_variables`, as we no longer assume that every lint
has a corresponding documentation URL and instead only add a URL if
Clippy thinks it exists.

Also, slightly simplify the test utility code for writing fake Clippy
results to a JSON-lines file.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645461
Original-Revision: 91820254a0e0ef03865e8df3f0f75ea92653e58b
GitOrigin-RevId: 1dad9b8561f7ef60bdbf77c88a444cc26304dfea
Change-Id: I78af58d5d1605ed9b77f4421f60ecf7ec0aaf62d