[roll] Roll fuchsia [run-test-suite] Save shell output as a custom artifact
Saves the shell output produced for each test run to custom artifacts.
Another option is to forward the stdout directly to the test case's
stdout, but this is likely to be very confusing when viewing the output.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597252
Original-Revision: 91bdf1cb2f7e75bd9ca91cae52afe1b0f5ca74f1
GitOrigin-RevId: b7cb3ddec08fb56825b4eb14bc0b62c54b218774
Change-Id: Ie98595a8e5b7486420a2c24e106648190068b75d