[roll] Roll fuchsia [phys] Introduce HandoffEntropy
This class job is to be the place where entropy sources from phys
are consumed. This involves copying out bytes and shredding
the contents.

For the zbi item the contents are zeroed after consumption.
For the cmdline entropy the contents are '0' after consumption.

Added zbi-handoff-entropy-tests to verify this behavior.

This CL does NOT introduce behavior changes with respect to the
existing mechanisms. That said, it is still not plugged in,
and that is reserved for a following change.

At this moment we dont have the mechanisms for exception
testing in phys, but we manually verified the behavior
with invalid inputs(invalid cmdline value, trying to
take entropy pool without enough entropy, etc).

This is in addition to existing emulator tests, that
already verified boot behavior under these scenarios.

Test: fx run-zbi-test qemu-phys-unittests.multiboot
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612417
Original-Revision: c876f8c722dd3f81eb3caba2a46887628017500e
GitOrigin-RevId: 97a6128ffb195d6d699fdd617e52f4df8970d7b6
Change-Id: I27df9bad140ed175f60ebb99e1b73015ac3a002b