[roll] Roll fuchsia [fs_management] Add support for matching children of a device.
The paver uses fs_management to find all FVM partitions of a particular
type so that it can wipe them before writing a new one to the FVM.
Currently, it doesn't support systems with more than one FVM - since
PartitionMatches always returns the first matching partition it finds,
it may return a partition belonging to the wrong FVM.

This constraint was fine previously - zedboot does not have a ramdisk
FVM built-in - but for writing a completely new FVM from recovery, we
need to handle the case where multiple FVMs are present on a system and
we need to distinguish between them.

This CL adds support for a "parent_device" member in `struct
PartitionMatcher`, which causes fs_management to check that the
found partition is a child of the given device.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645142
Original-Revision: a1e8d11e63a43dd4f9e323155574c56307293e1c
GitOrigin-RevId: f3b28b88e66bffc3851b1a949bdf854f1860425e
Change-Id: I04911040b5c6c8a4b75e66da7e5ca2bd8d36c43a