[roll] Roll fuchsia [mdns] Forward logs to serial console
Previously the mDNS component's std{out,err} was routed to klog and thus
ended up in the serial console; it inherited this behavior because it
ran in the sys realm under appmgr. However, after migrating to v2 in
commit e2b912dd5bfe025021d7db120a509e8ab6266d34, it lives under core, so
for logs to show up in the serial console, they must be explicitly
forwarded.

Original-Bug: 90433
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645471
Original-Revision: 503703156afa6d79e1874a6661a425004e60c1c0
GitOrigin-RevId: afb93cbe3792501af0688fed3579ee6f0d319a97
Change-Id: I00e0b6b732ace5853343eb13e9233e1eb3eeb482