[roll] Roll infra/recipes [testsharder] Include stderr in summary_markdown on failure
testsharder may fail as a result of user error, e.g. if a CL author
tries to use a multiplier that matches too many tests. In this case it's
important that the error message is easy to find, so we should bubble it
up to the summary_markdown when the build fails.

I tried to do this 2 years ago (https://fxrev.dev/389280) after I
implemented the multiplier validation logic but ended up reverting the
CL because it wasn't working as intended. I now realize that was because
I was expecting the error message to be in testsharder's stdout, when
it's actually in stderr.

Also do a little bit of cleanup of unused parameters in the test_api.

Original-Fixed: 93427
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/645982
Original-Revision: c3955590e2ee4b9ddaeaf595e816a1412387287d
GitOrigin-RevId: e0df494c09f334b7d1e881774feab0befb3d0cef
Change-Id: Ie78c8d3721738f60d12597f43cfcf56edd3d183f