[roll] Roll fuchsia [flatland] Modify flatland_benchmarks_test component kill order.
fxb/91066 is a flake where a RootPresenter from a previous test is
apparently not cleaned up properly.  Try topologically rearranging
the order that the test kills components, so that the RootPresenter
isn't restarted, and thus doesn't muck up the current test.

Also add "a11y-manager" to the list of killed components, since it
seems likely to be the culprit that is causing "root_presenter" to
restart.

Original-Bug: 91066
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644036
Original-Revision: 63d8004e3368c33d8419ffdc5c7e4f8b8e405e16
GitOrigin-RevId: 4523f094ba3d52d7568d5a2870578d64e34bb0a2
Change-Id: I727a0cdcc53f4f5bc3faae17b98332f9b0d77bd4