[roll] Roll fuchsia [audio_core] Add a deferred log buffer
It is sometimes useful to maintain a trace of the last N events before
an event of interest. While this can be done with the tracing system,
it's currently not easy to dump trace events during a test failure or
before an FX_CHECK (typically trace events need to be read by an
external program).

To work around those problems, this CL adds a VERBOSE_LOG macro. This
macro has two modes: In eager mode (useful for debugging tests), this
reduces to an immediate FX_LOGS call. In deferred mode, log messages are
placed into a circular buffer which is not flushed to the actual FX_LOG
until a flush is explicitly requested, either via a call to
DumpVerboseLogs() or an FX_CHECK failure. The buffer is thread-local to
avoid locking -- we can revisit this decision later if the need arises.

Performance note: it's not possible to preallocate a std::ostringstream,
so adding to a circular log can result in allocations. AFAICT this is
already the case with FX_LOGS(X) statements. If this becomes a concern,
we can either wrap VERBOSE_LOG statements with `if constexpr()` or
implement a std::ostream that preallocates its buffer.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639484
Original-Revision: 974e3a4716c0af1320df4799c22a6f73c02faf66
GitOrigin-RevId: f724bf1e421ac26a35c95be9c3b648e4fd95457b
Change-Id: Ia7a5813e27e761226887fc1a65aafb1b07c121a8