[roll] Roll fuchsia [kernel][object] Clean up MSI object names
Some transitions since implementation of the original MSI design and
RFCs led to naming semantics that needed to be cleaned up, and object
usage that was unclear when libzx code was generated. For instance, FIDL
tools typically wanted to generate references to zx::msi_allocation,
since technically according to our sources that was correct despite
libzx settling on zx::msi. Ultimately, this was my fault and this is an
attempt to fix it.

MsiDispatcher -> MsiInterruptDispatcher
- Despite the slightly redundant name, the new name makes it clear that
  this is a derived InterruptDispatcher for use with MSIs.

MsiAllocationDispatcher -> MsiDispatcher
- From the perspective of userspace we simply have handles to msi
  blocks, and zx::msi. Having zx::msi and MsiAllocation Dispatcher was
  causing issues between generated FIDL bindings and handle types, so
  this simplifies it such that MsiDispatcher and MsiInterruptDispatchers
  represent everything necessary, and from the perspective of handle
  typing and libzx we simply treat MsiInterruptDispatchers as
  interrupts. For a similar reason, the MSII object type has been
  removed since it was never actually used and contributed to technical
  debt.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645101
Original-Revision: 68e696137f6ef7b567a0c5851b966260c40a8aee
GitOrigin-RevId: 2a38901d649d0ba97212108b9015122edf903419
Change-Id: Ie8c1a56cf3230bd47c6ca872ec3d4e2f50fbc091