[roll] Roll fuchsia [kernel][vm] Pausable VMAR enumerator
This commit refactors the callback based enumeration in
EnumerateChildInternalLocked into a stateful enumerator that yields
regions. Further, this stateful enumerator is extended to be able to
pause and resume and tolerate the vmar hierarchy changing. The pause
and resume functionality is not yet used beyond tests.

Original-Bug: 84616

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640902
Original-Revision: 7409e1ed02aab09df28bbca59dbc250299886f1e
GitOrigin-RevId: 02aa0a259d93e0bf8094d91cdcc523bb1037eda4
Change-Id: I0fb223e1acffeaca0df2dd265ab2a7dbeb0392fc