[roll] Roll fuchsia [ffx] Add config setting to enable the repository server
This unifies fx and fserve to use the ffx config setting
repository.server.mode to determine which repository server. This avoids
the risk that the ffx daemon and the fx/fserve wrappers are in sync with
which repository server to use.

In order to do this, it:

* The ffx repository daemon service starts the server if
  repository.server.mode is "ffx".
* fserve to use repository.server.mode instead of fserve.server-mode.
* fx switches from using the feature flag "legacy_serve" to using
  repository.server.mode.

Original-Fixed: 93547

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645684
Original-Revision: 027beb6199a11f069957e2c5bd5d9cf1cc13e41e
GitOrigin-RevId: ea07b697c84321a7c9a8e2480f6bcc3ba24f5042
Change-Id: I92de59132bbb80525a545af7a4b7a63c4e00d511