[roll] Roll fuchsia [tools][fuzz] Improve QEMU boot monitoring logic
When starting a new instance, the previous code relied on `ps` to
determine whether QEMU had terminated early. However, because undercoat
detaches from QEMU and does not wait on the subprocess, `ps` will still
report an exited QEMU as a defunct process. This resulted in a deadlock
condition when a boot timed out. Instead of attempting to work around
this, which is OS-specific, this CL refactors the boot monitoring logic
to make it unnecessary.

Another benefit of the change is that some testing-specific `ps`
proxying code, which in fact masked the presence of this error, has been
removed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644033
Original-Revision: 00fc9296ea7a681b6a2c65460e6426e440fee5cf
GitOrigin-RevId: 12e6b4e16d907683aaafc3140ac4c83d03155c7b
Change-Id: I0b9b53006a9a59d4b8bfd627b8e37f455d6c4451