[roll] Roll fuchsia [ffx][test] Add missing artifact type to schema
Adds a missing artifact type to the schema.
This also adds a macro to autogenerate a list of all the variants for
an enum containing no fields, and uses the list to test all variants
against the schema. The macro is loosely based on a 'decodable_enum'
macro defined in //src/connectivity. This should at least block
submission when a variant not recognized in the schemas is added.

Testing: new unit tests, manually verify that adding a new variant
without updating schemas breaks tests.

Original-Bug: 81195

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646590
Original-Revision: 159165721d7adcbc97ba66c6013d0dbce1f7fed2
GitOrigin-RevId: 7d9b436f70faf84218eac65fe04032e7d0ed5987
Change-Id: I8fc10061c106ba0f7575aa991301b48ba51bc689