[roll] Roll fuchsia [ffx] Reuse image files for persisted emulators
The ffx emu shutdown command has a "--persist" flag, which leaves the
working directory in place after the emulator is terminated. These files
could be reused on future instances, but aren't currently. The "start"
command always re-stages the files, so the persisted versions are
overwritten.

This CL adds a check to the start-up code, such that if image files
already exist in the working directory, they aren't overwritten. The
upshot is that an emulator that was shutdown with the --persist flag can
effectively be "restarted", just by using the same instance name when
calling ffx emu start.

Original-Bug: fxbug.dev/80063
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645569
Original-Revision: cefe0d4bf7fd1b67282149c37702f33e86b67061
GitOrigin-RevId: b7bd4da163078c5d273729283c29d9cde801144a
Change-Id: I9c30131a0dc3e850dc9f59ca578a5fa1abdb3cf5