[roll] Roll llvm-project/libc [libc] change ASAN condition to generator expression
Previously, building LLVM-libc with GWP ASAN was conditioned on the flag
COMPILER_RT_BUILD_GWP_ASAN, which caused issues do to the default value
of the flag being set in the compiler-rt cmake, which is seperate. Now
GWP ASAN is included based on if it exists as a target, which is more
consistent.

Reviewed By: sivachandra

Differential Revision: https://reviews.llvm.org/D119789

GitOrigin-RevId: e623435fc1914d0f0253468c18e0bf13ca71811c
Original-Revision: a9039f8fde48a9ef0ce1db5a95ef18f0c93552b1
Change-Id: I28fab7073cc232a9978b178976131fc7a9877369