[roll] Roll fuchsia [build][python] Ability to specify the field serializer fn
ints with a value of 0 were not being serialized as part of the check
to see if a value was present or not.  This makes the logic detect
that situation, and adds the ability to specify a serializer fn with
a class decorator: '@serialize_field_as(my_field=str)'

Also adds the '@serialize_json' and '@serialize_dict' decorators that
I had previously written, but left in a stash.

Includes tests for all the various code-paths.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641711
Original-Revision: 67caac247585485567785483f034f27e91b693a7
GitOrigin-RevId: d0726153dc439f43bcb1b81ffd47b6606e7d91f7
Change-Id: I19a8f4ec024a1f10b9fffdceeee9a0af33d35dc0