[roll] Roll fuchsia [netcfg] Reject all unknown configurations
The `#[serde(deny_unknown_fields)]` attribute is set on the `Config`
type, but is not present on any of its members, though these are all
strict. This CL sets the attribute on all contained structs and enums
such that unrecognized options will be rejected anywhere in the
configuration.

It may also be useful to have runtime tests for this behavior in case
future changes forget to add this attribute. But that's outside the
scope of this CL.

Original-Bug: 93845
Test: added new test validating all types contained within `Config`.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646767
Original-Revision: af9ae385f8e100b1cb1eeb2c7a018ed7173725d6
GitOrigin-RevId: 5b9c4d7d5c918fce78243551bd7bd480e8496640
Change-Id: Idbe41100fabf72c507f5b30eb682f4d1e5188232