[roll] Roll fuchsia [cm] Introduce shutdown::Component trait.
Before this change, much of the shutdown logic gets its data about the
component that's shutting down from a `ComponentDecl`. This strategy
precludes using information like dynamic offers in computing the
shutdown order, and also makes it so that the tests for
`process_component_dependencies` can't know anything about dynamic
instances.

Behaviorally, this change is a no-op, but it replaces that use of
`ComponentDecl` with use of a new `shutdown::Component` trait. This
layer of indirection will allow us to account for dynamic offers and
children in subsequent CLs.

Original-Bug: 84678

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646702
Original-Revision: 9a56262e2795bac5fab587ee8bcc79ea94d21b82
GitOrigin-RevId: 138e9ccbb29ad81204edbb5bb926acfb413357c9
Change-Id: I3e914984456c71c7a6c82adc9c797cfb0c12f136