[roll] Roll fuchsia [devshell] fx format-code should not fail on one unknown extension
A fairly recent change was made to fail if `fx format-code` is passed a
"single file" with an unknown extension, but for multiple files, it was
supposed to format the known extensions and fail silently on unknown
extensions.

This change didn't work as intended because the check for multiple files
assumed the `FILES` variable was an array.

Usually, it was not an array, so the number of elements was never
greater than 1, despite the number of files.

It turns out that `FILES` is set in multiple places, and most don't make
it an array. (One does.) And many other places that used `FILES` used it
assuming it was an array. This only worked (most of the time) because
the spaces were expanded into multiple words upon evaluation.

In general, I'm surprised more things weren't breaking.

I updated all uses (set and evaluation) to use the array syntax, and
updated the test to validate that, with this change, now a list of files
with some known and some unknown extensions will still format the known
ones.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647567
Original-Revision: 3c40ccde4760b99468f27699ba7345560a5e5d93
GitOrigin-RevId: 57d35bdcb193dd02d0cd08a8b22b210ffae9ffdb
Change-Id: Ie43bcac9af942fa36c24de7d875a4b0834833962