[roll] Roll fuchsia [ffx][product] Remove workaround for bug 93799
The "start_up_args_template": "emulator_flags.json.template",
is now in the cb container. The workaround is not necessary
for current builds.

This workaround could be kept longer to support prior builds,
but there's not many, so it seems reasonable to remove this
workaround.

Original-Bug: 93799
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647662
Original-Revision: 11876ff259319c088b73f655ab754acc914194e2
GitOrigin-RevId: 66df11089dcc433e5f9a980cf071f05eb6e712a3
Change-Id: I71671940154b822b3e56f6697dc74c4aa026d098