[roll] Roll fuchsia [pwauth] Bump timeout in integration tests
On ASAN builds, we've observed this flake occur from time to time where
the devmgr never manages to bring up ramctl within 20 seconds.  We only
see this flake on the arm64-asan builder.

We have reason to believe that these integration tests are simply
running out of CPU time, due to the fact that many of them are doing the
work of full password verification, which is intentionally
computationally expensive to limit the rate at which attackers can
attempt to crack passwords.

To attempt to patch around the high flake rate, try increasing the
timeout, which should allow slightly more core time on slow ASAN
builders that are running VMs.  Perhaps in the fullness of time, we'll
move these tests to only run on real hardware, but we'll want adequate
CQ coverage if we go that route.

Original-Bug: 90826
Original-Bug: 93675
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647347
Original-Revision: 74704782c63c121b99549a44490472578764a584
GitOrigin-RevId: 97ca20fd2c399a46a47566b0fa48a634c4f5e6eb
Change-Id: Ic86e7a98c363c2174d7c9a7fd47ad60c17e48f12