[roll] Roll fuchsia [ffx] Change default behavior for name parameters
The attached bug indicated that stop and show should default to the
only running emulator when there is only one running, rather than having
a fixed default value. This simplifies the workflow for named emulators,
even if the default name is not used at start.

This CL modifies the behavior of both stop and show, such that if no
name is provided, the command will call get_all_instances, then if
only one value is returned that value will be used for the command. If
no instances are found, a message will be printed to indicate this to
the user, and the command will terminate. If multiple emulators are
running, a message asking the user to utilize the name parameter is
printed, and the command terminates. We also add return codes for both
subcommands, so that test infra can differentiate between the various
failure modes.

Original-Bug: fxbug.dev/93472
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647384
Original-Revision: 5352bbf4364ffe0cf13839d76a063eddf23bb95f
GitOrigin-RevId: 69a66a9f695a77897de59590e198615102f550ed
Change-Id: I012d8ec74abdf0ebce5239ea83b06e7fc4f8ee81