[roll] Roll fuchsia [tests][oom] Fix test race where we wait for a wrong message
This change fixes a bug in the oom_test that was introduced with the
kernel.oom.trigger-on-alloc-failure feature.  Depending on the exact
nature of the OOM event we may or may not observe the memory pressure
state change message.

Test: Ran host_x64/oom_tests 10 times.
Original-Fixed: 94019
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647801
Original-Revision: bcfeda7737f738960b2cbb14691c1ab089d2d78c
GitOrigin-RevId: 770a552781fe611b84fdf4695c008e0d6e5bf313
Change-Id: I44f05182dba523a0b14b4e8b8b4ef58bbc0876b3