[roll] Roll fuchsia [component_manager] fix DestroyChild race
The DestroyChild action was using non-instanced monikers. This means
that if a DestroyChild action was racing with a DestroyChild and a
CreateChild action, in rare cases the duplicate DestroyChild action
could destroy the instance created by the CreateChild instance.

To fix this, refactor the DestroyChild action to work with
InstancedChildMoniker, and have it check the instance of the child
returned by `get_live_child` to confirm that it's operating on the
correct child.

Generally actions that perform work on children should use instanced
monikers to reference those children, to remove
time-of-check-time-of-use issues where the live children set is changed
between when the action is created and when it's able to be
registered/run.

Original-Fixed: 92354
MULTIPLY: network-test-realm-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645942
Original-Revision: ea4ceefb6254de4b2913f9b44d105a6e5d2d5615
GitOrigin-RevId: 7cd7787a1bd95483958e17976dd9d781e97d8d92
Change-Id: I73aba1be62c60e00f13a150e3de75e149ebfe819