[roll] Roll experiences [session] Host tmp memfs for chrome in ermine
Chrome now uses `tmp` storage defined in the `ermine` shell component,
which is hosted by a sibling component, `memfs`.

This ensures that the contents of `tmp` are cleared when memfs is
stopped. Ermine must be stopped recursively for this to happen
on logout.

Original-Bug: 92782
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/647841
Original-Revision: 2748330c12a307d7561df2fff10b76cc7d1715c5
GitOrigin-RevId: be540d5c2d56cf66d99ca5bf3d7c32df41126512
Change-Id: Ib7e7b48026a1b11d5637f48c0c47d04775d0543b