[roll] Roll fuchsia [carnelian] default to scenes for rendering
Add a new method to the ViewAssistant trait to return an optional
mutable reference to the view's scene. Components which wish to use
scenes can implement this method to remove some boilerplate code.

In the future this will be used to allow user input events to be
delegated to scenes without requiring the view assistant to do the
forwarding. By this methods things like drag and drop can be easily
implemented in the scene.

As part of this CL I moved the button, font_metrics, layout_gallery and
spinning_square examples to take advantage of the new trait method.
Other scene-using samples were not using layout, so could not be easily
migrated since the new trait method calls layout, which fails when
there's no root group. This area needs to be cleaned up a bit then
everything scene-using can be moved over.

Testing: manual and existing unit tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638968
Original-Revision: 0d7dcedc50317f43a55ba688d1a24994464799b3
GitOrigin-RevId: e443c26443bf53dd09e56131797de464cbf3618b
Change-Id: Ie466a2657dc13623b3316722504845dccfb2259b