[roll] Roll llvm-project/libc [libc] Add exit and atexit
Often atexit is implemented using __cxa_atexit. I have not implemented __cxa_atexit here because it potentially requires more discussion. It is unique for llvm-libc (I think) that it is an exported symbol that wouldn’t be defined in any spec file because it doesn’t have a header. Implementing it will be trivial given what is here already, but I figured it would be more contentious so it can be implemented later.

Reviewed By: lntue

Differential Revision: https://reviews.llvm.org/D119512

GitOrigin-RevId: a011ea1646a7859996bcae2bf468d8962562bc41
Original-Revision: 6e0568b7f7221bd750135ba2a4d1c64194e86e3d
Change-Id: I0d1cd9fb127872b7f93f80b10a3e97eda6e78f50