[roll] Roll fuchsia [power] Create SystemPowerModeHandler node
Creates a new node called SystemPowerModeHandler. This node is
responsible for serving three new protocols:

  fuchsia.power.clientlevel.Connector - allows a client service to
  connect a fuchsia.power.clientlevel.Watcher server end to receive
  power level changes of a desired fuchsia.power.clientlevel.ClientType

  fuchsia.power.systemmode.Requester - allows a component to request
  changes to the currently active system power modes

  fuchsia.power.systemmode.ClientConfigurator - allows a component to
  reconfigure the system power mode configuration for a client

The node reads in the system power mode configuration file during
initialization and uses this config to determine the appropriate power
level for a client based on the currently active system power modes.

This CL simply adds the code for the new node, but it is not
instantiated on any product (yet).

Original-Bug: 91839
Test: fx test power-manager-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645685
Original-Revision: 5a9aba5090f985ceced9e67335325fd446ecaede
GitOrigin-RevId: e52c9e7c65fa51e20dcbd25faa064d501e0994aa
Change-Id: I94e76ec9b66a65c426b70a089a4bfbc379c9eae5