[roll] Roll fuchsia [sherlock][power] Do not reconfigure VDDEE_800 rail on init.
As part of PWM initialization we reconfigure all PWMs including
PWMAO_B which controls the VDDEE_800 rail which is already
configured by the bootloader. We should be leaving this rail as-is.

Setting the `protect` flag on the PWM prevents our PWM driver from
reconfiguring this rail.

Test: Smoke test completed by booting on a device
umangshah@ also verified the rail using a scope.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648004
Original-Revision: 66506790508aad289ea1fd53a2da51b63946e023
GitOrigin-RevId: cf925690f6cf81287a33b8d8b7a61f63c741c81e
Change-Id: Icfa2c48759aec1794b636145f3ef71332ca0fa84