[roll] Roll fuchsia [dart] Enable Dart API checking
The script "detect_api_changes" will now be run on all dart libraries
that are included in the SDK.

If any changes are detected, an error will
appear in CQ instructing the user to update the associated .api file.

Then, with a change to the .api file, an API reviewer will be required
to submit the change.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644269
Original-Revision: e279fd98f183ed7782f7d058afef97ff076ea19f
GitOrigin-RevId: c7575d3cd4da7e80143ffaa7eac9a341192090dc
Change-Id: I64c1f5d86b63f071b11c5cecf993e6d655e6fdbf