[roll] Roll fuchsia [test_list_tool] Derive tags from package manifests
Teaches the test_list_tool to add tags to TestListEntry based on the
contents of test package manifests.

Note that this CL does not correctly handle setting CFv1 or allowlisted
test packages to be non-hermetic. This will be addressed in a follow-up
CL.

Original-Bug: 92959
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648012
Original-Revision: c069cdcf435296cfacdcedf1ffba338958e61987
GitOrigin-RevId: 6c45a3c195671beefb3cdd5943bc1e2447938024
Change-Id: I535ac164dfb5d6777e787ea84d39d31df518f56b