[roll] Roll fuchsia [kernel][vm] Require PageRequest
Enforces the requirement of GetPage that a PageRequest be passed in if
fault flags are set, removing the temporary carveout for anonymous
VMOs.

Also annotates these locations that now need to pass in pagerequests as
needing validation that locks are not held when those pageqrequests are
in use.

Original-Bug: 94078

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647866
Original-Revision: e498024753d48bf34f93c4a4b891f6326c8df237
GitOrigin-RevId: 1e3fc20c5c7954f113fc1a3aebe857ae7535e2d3
Change-Id: Ib51011e9582441bca92af0a882eba4051dcfb401