[roll] Roll infra/recipes [qemu] Add support for building upstream HEAD
This change adds support for building upstream HEAD, while preserving
that for the fork. This was tested with QEMU recipe try jobs[1] as well
as locally in boot Fuchsia (which it does just fine).

An "experimental" tag is added if we upload to CIPD in the upstream
case. This will allow us test the new QEMU prebuilts against ~all
Fuchsia tests in dry-running a change in //integration to point to the
experimental version of the package.

[1] https://chromium-swarm.appspot.com/task?id=5923119d8faf7311&o=true&w=true

Original-Bug: 94023
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/647507
Original-Revision: dd7b6672453e76b6eb4cb67c5e6801585312e9d2
GitOrigin-RevId: 2d70ead66b36d0d2d5b4ddad931ad74ca34545e0
Change-Id: I6e5693ae504daae8f2074da1f4136721fa3e20b9